Update README, move craft code to seperate file/class
