+++ sources/craft.cpp
@@ -1,0 +1,3071 @@
+/*
+Craft engine originally written in C99
+Ported to C++11
+*/
+
+#include "craft.h"
+
+#include <dearimgui/imgui.h>
+#include <dearimgui/backends/imgui_impl_sdl3.h>
+#include <dearimgui/backends/imgui_impl_opengl3.h>
+
+#include <SDL3/SDL.h>
+
+#if defined(IMGUI_IMPL_OPENGL_ES2)
+#include <SDL3/SDL_opengles2.h>
+#else
+#include <SDL3/SDL_opengl.h>
+#endif
+#define SDL_FUNCTION_POINTER_IS_VOID_POINTER
+
+#include <curl/curl.h>
+
+#include <cmath>
+#include <cstdio>
+#include <cstdlib>
+#include <cstring>
+#include <ctime>
+#include <string>
+#include "auth.h"
+#include "client.h"
+#include "config.h"
+#include "cube.h"
+#include "db.h"
+#include "item.h"
+
+#include "matrix.h"
+#include <noise/noise.h>
+#include "util.h"
+#include "world.h"
+
+#define DUMP_GL_EXTENSIONS 0
+
+#define ALIGN_LEFT 0
+#define ALIGN_CENTER 1
+#define ALIGN_RIGHT 2
+
+#define MODE_OFFLINE 0
+#define MODE_ONLINE 1
+
+#define WORKER_IDLE 0
+#define WORKER_BUSY 1
+#define WORKER_DONE 2
+
+std::unique_ptr<craft::Model> craft::g (new craft::Model());
+
+/**
+ * @param seed defaults to zero
+*/
+craft::craft(const std::string& window_name, unsigned int seed)
+: window_name(window_name)
+, seed(seed) {
+
+}
+
+int craft::chunked(float x) const {
+    return SDL_floorf(SDL_roundf(x) / CHUNK_SIZE);
+}
+
+double craft::get_time() const {
+	return (SDL_GetTicks() + (double) g->start_time - (double) g->start_ticks) / 1000.0;
+}
+
+float craft::time_of_day() const {
+    if (g->day_length <= 0) {
+        return 0.5;
+    }
+    float t;
+    t = get_time();
+    t = t / g->day_length;
+    t = t - (int)t;
+    return t;
+}
+
+float craft::get_daylight() const {
+    float timer = time_of_day();
+    if (timer < 0.5) {
+        float t = (timer - 0.25) * 100;
+        return 1 / (1 + SDL_powf(2, -t));
+    }
+    else {
+        float t = (timer - 0.85) * 100;
+        return 1 - 1 / (1 + SDL_powf(2, -t));
+    }
+}
+
+int craft::get_scale_factor() const {
+    int window_width, window_height;
+    int buffer_width, buffer_height;
+    SDL_GetWindowSize(g->window, &window_width, &window_height);
+    SDL_GetWindowSizeInPixels(g->window, &buffer_width, &buffer_height);
+    int result = buffer_width / window_width;
+    result = MAX(1, result);
+    result = MIN(2, result);
+    return result;
+}
+
+void craft::get_sight_vector(float rx, float ry, float *vx, float *vy, float *vz) const {
+    float m = SDL_cosf(ry);
+    *vx = SDL_cosf(rx - RADIANS(90)) * m;
+    *vy = SDL_sinf(ry);
+    *vz = SDL_sinf(rx - RADIANS(90)) * m;
+}
+
+void craft::get_motion_vector(int flying, int sz, int sx, float rx, float ry,
+    float *vx, float *vy, float *vz) const {
+    *vx = 0; *vy = 0; *vz = 0;
+    if (!sz && !sx) {
+        return;
+    }
+    float strafe = SDL_atan2f(sz, sx);
+    if (flying) {
+        float m = SDL_cosf(ry);
+        float y = SDL_sinf(ry);
+        if (sx) {
+            if (!sz) {
+                y = 0;
+            }
+            m = 1;
+        }
+        if (sz > 0) {
+            y = -y;
+        }
+        *vx = SDL_cosf(rx + strafe) * m;
+        *vy = y;
+        *vz = SDL_sinf(rx + strafe) * m;
+    }
+    else {
+        *vx = SDL_cosf(rx + strafe);
+        *vy = 0;
+        *vz = SDL_sinf(rx + strafe);
+    }
+}
+
+GLuint craft::gen_crosshair_buffer() {
+    float x = static_cast<float>(g->width) / 2.0f;
+    float y = static_cast<float>(g->height) / 2.0f;
+    float p = 10.f * static_cast<float>(g->scale);
+    float data[] = {
+        x, y - p, x, y + p,
+        x - p, y, x + p, y
+    };
+    return gen_buffer(sizeof(data), data);
+}
+
+GLuint craft::gen_wireframe_buffer(float x, float y, float z, float n) {
+    float data[72];
+    make_cube_wireframe(data, x, y, z, n);
+    return gen_buffer(sizeof(data), data);
+}
+
+GLuint craft::gen_sky_buffer() {
+    float data[12288];
+    make_sphere(data, 1, 3);
+    return gen_buffer(sizeof(data), data);
+}
+
+GLuint craft::gen_cube_buffer(float x, float y, float z, float n, int w) {
+    GLfloat *data = malloc_faces(10, 6);
+    float ao[6][4] = {0};
+    float light[6][4] = {
+        {0.5, 0.5, 0.5, 0.5},
+        {0.5, 0.5, 0.5, 0.5},
+        {0.5, 0.5, 0.5, 0.5},
+        {0.5, 0.5, 0.5, 0.5},
+        {0.5, 0.5, 0.5, 0.5},
+        {0.5, 0.5, 0.5, 0.5}
+    };
+    make_cube(data, ao, light, 1, 1, 1, 1, 1, 1, x, y, z, n, w);
+    return gen_faces(10, 6, data);
+}
+
+GLuint craft::gen_plant_buffer(float x, float y, float z, float n, int w) {
+    GLfloat *data = malloc_faces(10, 4);
+    float ao = 0;
+    float light = 1;
+    make_plant(data, ao, light, x, y, z, n, w, 45);
+    return gen_faces(10, 4, data);
+}
+
+GLuint craft::gen_player_buffer(float x, float y, float z, float rx, float ry) {
+    GLfloat *data = malloc_faces(10, 6);
+    make_player(data, x, y, z, rx, ry);
+    return gen_faces(10, 6, data);
+}
+
+GLuint craft::gen_text_buffer(float x, float y, float n, char *text) {
+    int length = strlen(text);
+    GLfloat *data = malloc_faces(4, length);
+    for (int i = 0; i < length; i++) {
+        make_character(data + i * 24, x, y, n / 2, n, text[i]);
+        x += n;
+    }
+    return gen_faces(4, length, data);
+}
+
+void craft::draw_triangles_3d_ao(Attrib *attrib, GLuint buffer, int count) {
+    glBindBuffer(GL_ARRAY_BUFFER, buffer);
+    glEnableVertexAttribArray(attrib->position);
+    glEnableVertexAttribArray(attrib->normal);
+    glEnableVertexAttribArray(attrib->uv);
+    glVertexAttribPointer(attrib->position, 3, GL_FLOAT, GL_FALSE,
+        sizeof(GLfloat) * 10, 0);
+    glVertexAttribPointer(attrib->normal, 3, GL_FLOAT, GL_FALSE,
+        sizeof(GLfloat) * 10, (GLvoid *)(sizeof(GLfloat) * 3));
+    glVertexAttribPointer(attrib->uv, 4, GL_FLOAT, GL_FALSE,
+        sizeof(GLfloat) * 10, (GLvoid *)(sizeof(GLfloat) * 6));
+    glDrawArrays(GL_TRIANGLES, 0, count);
+    glDisableVertexAttribArray(attrib->position);
+    glDisableVertexAttribArray(attrib->normal);
+    glDisableVertexAttribArray(attrib->uv);
+    glBindBuffer(GL_ARRAY_BUFFER, 0);
+}
+
+void craft::draw_triangles_3d_text(Attrib *attrib, GLuint buffer, int count) {
+    glBindBuffer(GL_ARRAY_BUFFER, buffer);
+    glEnableVertexAttribArray(attrib->position);
+    glEnableVertexAttribArray(attrib->uv);
+    glVertexAttribPointer(attrib->position, 3, GL_FLOAT, GL_FALSE,
+        sizeof(GLfloat) * 5, 0);
+    glVertexAttribPointer(attrib->uv, 2, GL_FLOAT, GL_FALSE,
+        sizeof(GLfloat) * 5, (GLvoid *)(sizeof(GLfloat) * 3));
+    glDrawArrays(GL_TRIANGLES, 0, count);
+    glDisableVertexAttribArray(attrib->position);
+    glDisableVertexAttribArray(attrib->uv);
+    glBindBuffer(GL_ARRAY_BUFFER, 0);
+}
+
+void craft::draw_triangles_3d(Attrib *attrib, GLuint buffer, int count) {
+    
+    glBindBuffer(GL_ARRAY_BUFFER, buffer);
+    glEnableVertexAttribArray(attrib->position);
+    glEnableVertexAttribArray(attrib->normal);
+    glEnableVertexAttribArray(attrib->uv);
+    
+    glVertexAttribPointer(attrib->position, 3, GL_FLOAT, GL_FALSE,
+        sizeof(GLfloat) * 8, 0);
+    glVertexAttribPointer(attrib->normal, 3, GL_FLOAT, GL_FALSE,
+        sizeof(GLfloat) * 8, (GLvoid *)(sizeof(GLfloat) * 3));
+    glVertexAttribPointer(attrib->uv, 2, GL_FLOAT, GL_FALSE,
+        sizeof(GLfloat) * 8, (GLvoid *)(sizeof(GLfloat) * 6));
+    glDrawArrays(GL_TRIANGLES, 0, count);
+    glDisableVertexAttribArray(attrib->position);
+    glDisableVertexAttribArray(attrib->normal);
+    glDisableVertexAttribArray(attrib->uv);
+    glBindBuffer(GL_ARRAY_BUFFER, 0);
+}
+
+void craft::draw_triangles_2d(Attrib *attrib, GLuint buffer, int count) {
+    glBindBuffer(GL_ARRAY_BUFFER, buffer);
+    glEnableVertexAttribArray(attrib->position);
+    glEnableVertexAttribArray(attrib->uv);
+    glVertexAttribPointer(attrib->position, 2, GL_FLOAT, GL_FALSE,
+        sizeof(GLfloat) * 4, 0);
+    glVertexAttribPointer(attrib->uv, 2, GL_FLOAT, GL_FALSE,
+        sizeof(GLfloat) * 4, (GLvoid *)(sizeof(GLfloat) * 2));
+    glDrawArrays(GL_TRIANGLES, 0, count);
+    glDisableVertexAttribArray(attrib->position);
+    glDisableVertexAttribArray(attrib->uv);
+    glBindBuffer(GL_ARRAY_BUFFER, 0);
+}
+
+void craft::draw_lines(Attrib *attrib, GLuint buffer, int components, int count) {
+    glBindBuffer(GL_ARRAY_BUFFER, buffer);
+    glEnableVertexAttribArray(attrib->position);
+    glVertexAttribPointer(
+        attrib->position, components, GL_FLOAT, GL_FALSE, 0, 0);
+    glDrawArrays(GL_LINES, 0, count);
+    glDisableVertexAttribArray(attrib->position);
+    glBindBuffer(GL_ARRAY_BUFFER, 0);
+}
+
+void craft::draw_chunk(Attrib *attrib, Chunk *chunk) {
+    draw_triangles_3d_ao(attrib, chunk->buffer, chunk->faces * 6);
+}
+
+void craft::draw_item(Attrib *attrib, GLuint buffer, int count) {
+    draw_triangles_3d_ao(attrib, buffer, count);
+}
+
+void craft::draw_text(Attrib *attrib, GLuint buffer, int length) {
+    glEnable(GL_BLEND);
+    glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
+    draw_triangles_2d(attrib, buffer, length * 6);
+    glDisable(GL_BLEND);
+}
+
+void craft::draw_signs(Attrib *attrib, Chunk *chunk) {
+    glEnable(GL_POLYGON_OFFSET_FILL);
+    glPolygonOffset(-8, -1024);
+    draw_triangles_3d_text(attrib, chunk->sign_buffer, chunk->sign_faces * 6);
+    glDisable(GL_POLYGON_OFFSET_FILL);
+}
+
+void craft::draw_sign(Attrib *attrib, GLuint buffer, int length) {
+    glEnable(GL_POLYGON_OFFSET_FILL);
+    glPolygonOffset(-8, -1024);
+    draw_triangles_3d_text(attrib, buffer, length * 6);
+    glDisable(GL_POLYGON_OFFSET_FILL);
+}
+
+void craft::draw_cube(Attrib *attrib, GLuint buffer) {
+    draw_item(attrib, buffer, 36);
+}
+
+void craft::draw_plant(Attrib *attrib, GLuint buffer) {
+    draw_item(attrib, buffer, 24);
+}
+
+void craft::draw_player(Attrib *attrib, Player *player) {
+    draw_cube(attrib, player->buffer);
+}
+
+craft::Player *craft::find_player(int id) {
+    for (int i = 0; i < g->player_count; i++) {
+        Player *player = g->players + i;
+        if (player->id == id) {
+            return player;
+        }
+    }
+    return 0;
+}
+
+void craft::update_player(Player *player, float x, float y, float z, float rx, float ry, int interpolate)
+{
+    if (interpolate) {
+        State *s1 = &player->state1;
+        State *s2 = &player->state2;
+        memcpy(s1, s2, sizeof(State));
+        s2->x = x; s2->y = y; s2->z = z; s2->rx = rx; s2->ry = ry;
+        s2->t = get_time();
+        if (s2->rx - s1->rx > PI) {
+            s1->rx += 2 * PI;
+        }
+        if (s1->rx - s2->rx > PI) {
+            s1->rx -= 2 * PI;
+        }
+    }
+    else {
+        State *s = &player->state;
+        s->x = x; s->y = y; s->z = z; s->rx = rx; s->ry = ry;
+        del_buffer(player->buffer);
+        player->buffer = gen_player_buffer(s->x, s->y, s->z, s->rx, s->ry);
+    }
+}
+
+void craft::interpolate_player(Player *player) {
+    State *s1 = &player->state1;
+    State *s2 = &player->state2;
+    float t1 = s2->t - s1->t;
+    float t2 = get_time() - s2->t;
+    t1 = MIN(t1, 1);
+    t1 = MAX(t1, 0.1);
+    float p = MIN(t2 / t1, 1);
+    update_player(
+        player,
+        s1->x + (s2->x - s1->x) * p,
+        s1->y + (s2->y - s1->y) * p,
+        s1->z + (s2->z - s1->z) * p,
+        s1->rx + (s2->rx - s1->rx) * p,
+        s1->ry + (s2->ry - s1->ry) * p,
+        0);
+}
+
+void craft::delete_player(int id) {
+    Player *player = find_player(id);
+    if (!player) {
+        return;
+    }
+    int count = g->player_count;
+    del_buffer(player->buffer);
+    Player *other = g->players + (--count);
+    memcpy(player, other, sizeof(Player));
+    g->player_count = count;
+}
+
+void craft::delete_all_players() {
+    for (int i = 0; i < g->player_count; i++) {
+        Player *player = g->players + i;
+        del_buffer(player->buffer);
+    }
+    g->player_count = 0;
+}
+
+float craft::player_player_distance(Player *p1, Player *p2) {
+    State *s1 = &p1->state;
+    State *s2 = &p2->state;
+    float x = s2->x - s1->x;
+    float y = s2->y - s1->y;
+    float z = s2->z - s1->z;
+    return SDL_sqrt(x * x + y * y + z * z);
+}
+
+float craft::player_crosshair_distance(Player *p1, Player *p2) {
+    State *s1 = &p1->state;
+    State *s2 = &p2->state;
+    float d = player_player_distance(p1, p2);
+    float vx, vy, vz;
+    get_sight_vector(s1->rx, s1->ry, &vx, &vy, &vz);
+    vx *= d; vy *= d; vz *= d;
+    float px, py, pz;
+    px = s1->x + vx; py = s1->y + vy; pz = s1->z + vz;
+    float x = s2->x - px;
+    float y = s2->y - py;
+    float z = s2->z - pz;
+    return SDL_sqrt(x * x + y * y + z * z);
+}
+
+craft::Player *craft::player_crosshair(Player *player) {
+    Player *result = 0;
+    float threshold = RADIANS(5);
+    float best = 0;
+    for (int i = 0; i < g->player_count; i++) {
+        Player *other = g->players + i;
+        if (other == player) {
+            continue;
+        }
+        float p = player_crosshair_distance(player, other);
+        float d = player_player_distance(player, other);
+        if (d < 96 && p / d < threshold) {
+            if (best == 0 || d < best) {
+                best = d;
+                result = other;
+            }
+        }
+    }
+    return result;
+}
+
+craft::Chunk *craft::find_chunk(int p, int q) {
+    for (int i = 0; i < g->chunk_count; i++) {
+        Chunk *chunk = g->chunks + i;
+        if (chunk->p == p && chunk->q == q) {
+            return chunk;
+        }
+    }
+    return 0;
+}
+
+int craft::chunk_distance(Chunk *chunk, int p, int q) {
+    int dp = ABS(chunk->p - p);
+    int dq = ABS(chunk->q - q);
+    return MAX(dp, dq);
+}
+
+int craft::chunk_visible(float planes[6][4], int p, int q, int miny, int maxy) {
+    float miny_f = static_cast<float>(miny);
+    float maxy_f = static_cast<float>(maxy);
+    float x = static_cast<float>(p * CHUNK_SIZE - 1);
+    float z = static_cast<float>(q * CHUNK_SIZE - 1);
+    float d = static_cast<float>(CHUNK_SIZE + 1);
+    float points[8][3] = {
+        {x + 0.f, miny_f, z + 0.f},
+        {x + d, miny_f, z + 0.f},
+        {x + 0.f, miny_f, z + d},
+        {x + d, miny_f, z + d},
+        {x + 0.f, maxy_f, z + 0.f},
+        {x + d, maxy_f, z + 0.f},
+        {x + 0.f, maxy_f, z + d},
+        {x + d, maxy_f, z + d}
+    };
+    int n = g->ortho ? 4 : 6;
+    for (int i = 0; i < n; i++) {
+        int in = 0;
+        int out = 0;
+        for (int j = 0; j < 8; j++) {
+            float d =
+                planes[i][0] * points[j][0] +
+                planes[i][1] * points[j][1] +
+                planes[i][2] * points[j][2] +
+                planes[i][3];
+            if (d < 0) {
+                out++;
+            }
+            else {
+                in++;
+            }
+            if (in && out) {
+                break;
+            }
+        }
+        if (in == 0) {
+            return 0;
+        }
+    }
+    return 1;
+}
+
+int craft::highest_block(float x, float z) {
+    int result = -1;
+    int nx = SDL_roundf(x);
+    int nz = SDL_roundf(z);
+    int p = chunked(x);
+    int q = chunked(z);
+    Chunk *chunk = find_chunk(p, q);
+    if (chunk) {
+        Map *map = &chunk->map;
+        MAP_FOR_EACH(map, ex, ey, ez, ew) {
+            if (is_obstacle(ew) && ex == nx && ez == nz) {
+                result = MAX(result, ey);
+            }
+        } END_MAP_FOR_EACH;
+    }
+    return result;
+}
+
+int craft::_hit_test (
+    Map *map, float max_distance, int previous,
+    float x, float y, float z,
+    float vx, float vy, float vz,
+    int *hx, int *hy, int *hz)
+{
+    int m = 32;
+    int px = 0;
+    int py = 0;
+    int pz = 0;
+    for (int i = 0; i < max_distance * m; i++) {
+        int nx = SDL_roundf(x);
+        int ny = SDL_roundf(y);
+        int nz = SDL_roundf(z);
+        if (nx != px || ny != py || nz != pz) {
+            int hw = map_get(map, nx, ny, nz);
+            if (hw > 0) {
+                if (previous) {
+                    *hx = px; *hy = py; *hz = pz;
+                }
+                else {
+                    *hx = nx; *hy = ny; *hz = nz;
+                }
+                return hw;
+            }
+            px = nx; py = ny; pz = nz;
+        }
+        x += vx / m; y += vy / m; z += vz / m;
+    }
+    return 0;
+}
+
+int craft::hit_test(
+    int previous, float x, float y, float z, float rx, float ry,
+    int *bx, int *by, int *bz)
+{
+    int result = 0;
+    float best = 0;
+    int p = chunked(x);
+    int q = chunked(z);
+    float vx, vy, vz;
+    get_sight_vector(rx, ry, &vx, &vy, &vz);
+    for (int i = 0; i < g->chunk_count; i++) {
+        Chunk *chunk = g->chunks + i;
+        if (chunk_distance(chunk, p, q) > 1) {
+            continue;
+        }
+        int hx, hy, hz;
+        int hw = _hit_test(&chunk->map, 8, previous,
+            x, y, z, vx, vy, vz, &hx, &hy, &hz);
+        if (hw > 0) {
+            float d = SDL_sqrt(SDL_powf(hx - x, 2) + SDL_powf(hy - y, 2) + SDL_powf(hz - z, 2));
+            if (best == 0 || d < best) {
+                best = d;
+                *bx = hx; *by = hy; *bz = hz;
+                result = hw;
+            }
+        }
+    }
+    return result;
+}
+
+int craft::hit_test_face(Player *player, int *x, int *y, int *z, int *face) {
+    State *s = &player->state;
+    int w = hit_test(0, s->x, s->y, s->z, s->rx, s->ry, x, y, z);
+    if (is_obstacle(w)) {
+        int hx, hy, hz;
+        hit_test(1, s->x, s->y, s->z, s->rx, s->ry, &hx, &hy, &hz);
+        int dx = hx - *x;
+        int dy = hy - *y;
+        int dz = hz - *z;
+        if (dx == -1 && dy == 0 && dz == 0) {
+            *face = 0; return 1;
+        }
+        if (dx == 1 && dy == 0 && dz == 0) {
+            *face = 1; return 1;
+        }
+        if (dx == 0 && dy == 0 && dz == -1) {
+            *face = 2; return 1;
+        }
+        if (dx == 0 && dy == 0 && dz == 1) {
+            *face = 3; return 1;
+        }
+        if (dx == 0 && dy == 1 && dz == 0) {
+            int degrees = SDL_roundf(DEGREES(SDL_atan2f(s->x - hx, s->z - hz)));
+            if (degrees < 0) {
+                degrees += 360;
+            }
+            int top = ((degrees + 45) / 90) % 4;
+            *face = 4 + top; return 1;
+        }
+    }
+    return 0;
+}
+
+int craft::collide(int height, float *x, float *y, float *z) {
+    int result = 0;
+    int p = chunked(*x);
+    int q = chunked(*z);
+    Chunk *chunk = find_chunk(p, q);
+    if (!chunk) {
+        return result;
+    }
+    Map *map = &chunk->map;
+    int nx = SDL_roundf(*x);
+    int ny = SDL_roundf(*y);
+    int nz = SDL_roundf(*z);
+    float px = *x - nx;
+    float py = *y - ny;
+    float pz = *z - nz;
+    float pad = 0.25;
+    for (int dy = 0; dy < height; dy++) {
+        if (px < -pad && is_obstacle(map_get(map, nx - 1, ny - dy, nz))) {
+            *x = nx - pad;
+        }
+        if (px > pad && is_obstacle(map_get(map, nx + 1, ny - dy, nz))) {
+            *x = nx + pad;
+        }
+        if (py < -pad && is_obstacle(map_get(map, nx, ny - dy - 1, nz))) {
+            *y = ny - pad;
+            result = 1;
+        }
+        if (py > pad && is_obstacle(map_get(map, nx, ny - dy + 1, nz))) {
+            *y = ny + pad;
+            result = 1;
+        }
+        if (pz < -pad && is_obstacle(map_get(map, nx, ny - dy, nz - 1))) {
+            *z = nz - pad;
+        }
+        if (pz > pad && is_obstacle(map_get(map, nx, ny - dy, nz + 1))) {
+            *z = nz + pad;
+        }
+    }
+    return result;
+}
+
+int craft::player_intersects_block(
+    int height,
+    float x, float y, float z,
+    int hx, int hy, int hz)
+{
+    int nx = SDL_roundf(x);
+    int ny = SDL_roundf(y);
+    int nz = SDL_roundf(z);
+    for (int i = 0; i < height; i++) {
+        if (nx == hx && ny - i == hy && nz == hz) {
+            return 1;
+        }
+    }
+    return 0;
+}
+
+int craft::_gen_sign_buffer(
+    GLfloat *data, float x, float y, float z, int face, const char *text)
+{
+    static const int glyph_dx[8] = {0, 0, -1, 1, 1, 0, -1, 0};
+    static const int glyph_dz[8] = {1, -1, 0, 0, 0, -1, 0, 1};
+    static const int line_dx[8] = {0, 0, 0, 0, 0, 1, 0, -1};
+    static const int line_dy[8] = {-1, -1, -1, -1, 0, 0, 0, 0};
+    static const int line_dz[8] = {0, 0, 0, 0, 1, 0, -1, 0};
+    if (face < 0 || face >= 8) {
+        return 0;
+    }
+    int count = 0;
+    float max_width = 64;
+    float line_height = 1.25;
+    char lines[1024];
+    int rows = wrap(text, max_width, lines, 1024);
+    rows = MIN(rows, 5);
+    int dx = glyph_dx[face];
+    int dz = glyph_dz[face];
+    int ldx = line_dx[face];
+    int ldy = line_dy[face];
+    int ldz = line_dz[face];
+    float n = 1.0 / (max_width / 10);
+    float sx = x - n * (rows - 1) * (line_height / 2) * ldx;
+    float sy = y - n * (rows - 1) * (line_height / 2) * ldy;
+    float sz = z - n * (rows - 1) * (line_height / 2) * ldz;
+    char *key;
+    char *line = tokenize(lines, "\n", &key);
+    while (line) {
+        int length = strlen(line);
+        int line_width = string_width(line);
+        line_width = MIN(line_width, max_width);
+        float rx = sx - dx * line_width / max_width / 2;
+        float ry = sy;
+        float rz = sz - dz * line_width / max_width / 2;
+        for (int i = 0; i < length; i++) {
+            int width = char_width(line[i]);
+            line_width -= width;
+            if (line_width < 0) {
+                break;
+            }
+            rx += dx * width / max_width / 2;
+            rz += dz * width / max_width / 2;
+            if (line[i] != ' ') {
+                make_character_3d(
+                    data + count * 30, rx, ry, rz, n / 2, face, line[i]);
+                count++;
+            }
+            rx += dx * width / max_width / 2;
+            rz += dz * width / max_width / 2;
+        }
+        sx += n * line_height * ldx;
+        sy += n * line_height * ldy;
+        sz += n * line_height * ldz;
+        line = tokenize(NULL, "\n", &key);
+        rows--;
+        if (rows <= 0) {
+            break;
+        }
+    }
+    return count;
+}
+
+void craft::gen_sign_buffer(Chunk *chunk) {
+    SignList *signs = &chunk->signs;
+
+    // first pass - count characters
+    int max_faces = 0;
+    for (int i = 0; i < signs->size; i++) {
+        Sign *e = signs->data + i;
+        max_faces += strlen(e->text);
+    }
+
+    // second pass - generate geometry
+    GLfloat *data = malloc_faces(5, max_faces);
+    int faces = 0;
+    for (int i = 0; i < signs->size; i++) {
+        Sign *e = signs->data + i;
+        faces += _gen_sign_buffer(
+            data + faces * 30, e->x, e->y, e->z, e->face, e->text);
+    }
+
+    del_buffer(chunk->sign_buffer);
+    chunk->sign_buffer = gen_faces(5, faces, data);
+    chunk->sign_faces = faces;
+}
+
+int craft::has_lights(Chunk *chunk) {
+    if (!SHOW_LIGHTS) {
+        return 0;
+    }
+    for (int dp = -1; dp <= 1; dp++) {
+        for (int dq = -1; dq <= 1; dq++) {
+            Chunk *other = chunk;
+            if (dp || dq) {
+                other = find_chunk(chunk->p + dp, chunk->q + dq);
+            }
+            if (!other) {
+                continue;
+            }
+            Map *map = &other->lights;
+            if (map->size) {
+                return 1;
+            }
+        }
+    }
+    return 0;
+}
+
+void craft::dirty_chunk(Chunk *chunk) {
+    chunk->dirty = 1;
+    if (has_lights(chunk)) {
+        for (int dp = -1; dp <= 1; dp++) {
+            for (int dq = -1; dq <= 1; dq++) {
+                Chunk *other = find_chunk(chunk->p + dp, chunk->q + dq);
+                if (other) {
+                    other->dirty = 1;
+                }
+            }
+        }
+    }
+}
+
+void craft::occlusion(
+    char neighbors[27], char lights[27], float shades[27],
+    float ao[6][4], float light[6][4])
+{
+    static const int lookup3[6][4][3] = {
+        {{0, 1, 3}, {2, 1, 5}, {6, 3, 7}, {8, 5, 7}},
+        {{18, 19, 21}, {20, 19, 23}, {24, 21, 25}, {26, 23, 25}},
+        {{6, 7, 15}, {8, 7, 17}, {24, 15, 25}, {26, 17, 25}},
+        {{0, 1, 9}, {2, 1, 11}, {18, 9, 19}, {20, 11, 19}},
+        {{0, 3, 9}, {6, 3, 15}, {18, 9, 21}, {24, 15, 21}},
+        {{2, 5, 11}, {8, 5, 17}, {20, 11, 23}, {26, 17, 23}}
+    };
+   static const int lookup4[6][4][4] = {
+        {{0, 1, 3, 4}, {1, 2, 4, 5}, {3, 4, 6, 7}, {4, 5, 7, 8}},
+        {{18, 19, 21, 22}, {19, 20, 22, 23}, {21, 22, 24, 25}, {22, 23, 25, 26}},
+        {{6, 7, 15, 16}, {7, 8, 16, 17}, {15, 16, 24, 25}, {16, 17, 25, 26}},
+        {{0, 1, 9, 10}, {1, 2, 10, 11}, {9, 10, 18, 19}, {10, 11, 19, 20}},
+        {{0, 3, 9, 12}, {3, 6, 12, 15}, {9, 12, 18, 21}, {12, 15, 21, 24}},
+        {{2, 5, 11, 14}, {5, 8, 14, 17}, {11, 14, 20, 23}, {14, 17, 23, 26}}
+    };
+    static const float curve[4] = {0.0, 0.25, 0.5, 0.75};
+    for (int i = 0; i < 6; i++) {
+        for (int j = 0; j < 4; j++) {
+            int corner = neighbors[lookup3[i][j][0]];
+            int side1 = neighbors[lookup3[i][j][1]];
+            int side2 = neighbors[lookup3[i][j][2]];
+            int value = side1 && side2 ? 3 : corner + side1 + side2;
+            float shade_sum = 0;
+            float light_sum = 0;
+            int is_light = lights[13] == 15;
+            for (int k = 0; k < 4; k++) {
+                shade_sum += shades[lookup4[i][j][k]];
+                light_sum += lights[lookup4[i][j][k]];
+            }
+            if (is_light) {
+                light_sum = 15 * 4 * 10;
+            }
+            float total = curve[value] + shade_sum / 4.0;
+            ao[i][j] = MIN(total, 1.0);
+            light[i][j] = light_sum / 15.0 / 4.0;
+        }
+    }
+}
+
+#define XZ_SIZE (CHUNK_SIZE * 3 + 2)
+#define XZ_LO (CHUNK_SIZE)
+#define XZ_HI (CHUNK_SIZE * 2 + 1)
+#define Y_SIZE 258
+#define XYZ(x, y, z) ((y) * XZ_SIZE * XZ_SIZE + (x) * XZ_SIZE + (z))
+#define XZ(x, z) ((x) * XZ_SIZE + (z))
+
+void craft::light_fill(
+    char *opaque, char *light,
+    int x, int y, int z, int w, int force)
+{
+    if (x + w < XZ_LO || z + w < XZ_LO) {
+        return;
+    }
+    if (x - w > XZ_HI || z - w > XZ_HI) {
+        return;
+    }
+    if (y < 0 || y >= Y_SIZE) {
+        return;
+    }
+    if (light[XYZ(x, y, z)] >= w) {
+        return;
+    }
+    if (!force && opaque[XYZ(x, y, z)]) {
+        return;
+    }
+    light[XYZ(x, y, z)] = w--;
+    light_fill(opaque, light, x - 1, y, z, w, 0);
+    light_fill(opaque, light, x + 1, y, z, w, 0);
+    light_fill(opaque, light, x, y - 1, z, w, 0);
+    light_fill(opaque, light, x, y + 1, z, w, 0);
+    light_fill(opaque, light, x, y, z - 1, w, 0);
+    light_fill(opaque, light, x, y, z + 1, w, 0);
+}
+
+void craft::compute_chunk(WorkerItem *item) {
+    char *opaque = (char *)calloc(XZ_SIZE * XZ_SIZE * Y_SIZE, sizeof(char));
+    char *light = (char *)calloc(XZ_SIZE * XZ_SIZE * Y_SIZE, sizeof(char));
+    char *highest = (char *)calloc(XZ_SIZE * XZ_SIZE, sizeof(char));
+
+    int ox = item->p * CHUNK_SIZE - CHUNK_SIZE - 1;
+    int oy = -1;
+    int oz = item->q * CHUNK_SIZE - CHUNK_SIZE - 1;
+
+    // check for lights
+    int has_light = 0;
+    if (SHOW_LIGHTS) {
+        for (int a = 0; a < 3; a++) {
+            for (int b = 0; b < 3; b++) {
+                Map *map = item->light_maps[a][b];
+                if (map && map->size) {
+                    has_light = 1;
+                }
+            }
+        }
+    }
+
+    // populate opaque array
+    for (int a = 0; a < 3; a++) {
+        for (int b = 0; b < 3; b++) {
+            Map *map = item->block_maps[a][b];
+            if (!map) {
+                continue;
+            }
+            MAP_FOR_EACH(map, ex, ey, ez, ew) {
+                int x = ex - ox;
+                int y = ey - oy;
+                int z = ez - oz;
+                int w = ew;
+                // TODO: this should be unnecessary
+                if (x < 0 || y < 0 || z < 0) {
+                    continue;
+                }
+                if (x >= XZ_SIZE || y >= Y_SIZE || z >= XZ_SIZE) {
+                    continue;
+                }
+                // END TODO
+                opaque[XYZ(x, y, z)] = !is_transparent(w);
+                if (opaque[XYZ(x, y, z)]) {
+                    highest[XZ(x, z)] = MAX(highest[XZ(x, z)], y);
+                }
+            } END_MAP_FOR_EACH;
+        }
+    }
+
+    // flood fill light intensities
+    if (has_light) {
+        for (int a = 0; a < 3; a++) {
+            for (int b = 0; b < 3; b++) {
+                Map *map = item->light_maps[a][b];
+                if (!map) {
+                    continue;
+                }
+                MAP_FOR_EACH(map, ex, ey, ez, ew) {
+                    int x = ex - ox;
+                    int y = ey - oy;
+                    int z = ez - oz;
+                    light_fill(opaque, light, x, y, z, ew, 1);
+                } END_MAP_FOR_EACH;
+            }
+        }
+    }
+
+    Map *map = item->block_maps[1][1];
+
+    // count exposed faces
+    int miny = 256;
+    int maxy = 0;
+    int faces = 0;
+    MAP_FOR_EACH(map, ex, ey, ez, ew) {
+        if (ew <= 0) {
+            continue;
+        }
+        int x = ex - ox;
+        int y = ey - oy;
+        int z = ez - oz;
+        int f1 = !opaque[XYZ(x - 1, y, z)];
+        int f2 = !opaque[XYZ(x + 1, y, z)];
+        int f3 = !opaque[XYZ(x, y + 1, z)];
+        int f4 = !opaque[XYZ(x, y - 1, z)] && (ey > 0);
+        int f5 = !opaque[XYZ(x, y, z - 1)];
+        int f6 = !opaque[XYZ(x, y, z + 1)];
+        int total = f1 + f2 + f3 + f4 + f5 + f6;
+        if (total == 0) {
+            continue;
+        }
+        if (is_plant(ew)) {
+            total = 4;
+        }
+        miny = MIN(miny, ey);
+        maxy = MAX(maxy, ey);
+        faces += total;
+    } END_MAP_FOR_EACH;
+
+    // generate geometry
+    GLfloat *data = malloc_faces(10, faces);
+    int offset = 0;
+    MAP_FOR_EACH(map, ex, ey, ez, ew) {
+        if (ew <= 0) {
+            continue;
+        }
+        int x = ex - ox;
+        int y = ey - oy;
+        int z = ez - oz;
+        int f1 = !opaque[XYZ(x - 1, y, z)];
+        int f2 = !opaque[XYZ(x + 1, y, z)];
+        int f3 = !opaque[XYZ(x, y + 1, z)];
+        int f4 = !opaque[XYZ(x, y - 1, z)] && (ey > 0);
+        int f5 = !opaque[XYZ(x, y, z - 1)];
+        int f6 = !opaque[XYZ(x, y, z + 1)];
+        int total = f1 + f2 + f3 + f4 + f5 + f6;
+        if (total == 0) {
+            continue;
+        }
+        char neighbors[27] = {0};
+        char lights[27] = {0};
+        float shades[27] = {0};
+        int index = 0;
+        for (int dx = -1; dx <= 1; dx++) {
+            for (int dy = -1; dy <= 1; dy++) {
+                for (int dz = -1; dz <= 1; dz++) {
+                    neighbors[index] = opaque[XYZ(x + dx, y + dy, z + dz)];
+                    lights[index] = light[XYZ(x + dx, y + dy, z + dz)];
+                    shades[index] = 0;
+                    if (y + dy <= highest[XZ(x + dx, z + dz)]) {
+                        for (int oy = 0; oy < 8; oy++) {
+                            if (opaque[XYZ(x + dx, y + dy + oy, z + dz)]) {
+                                shades[index] = 1.0 - oy * 0.125;
+                                break;
+                            }
+                        }
+                    }
+                    index++;
+                }
+            }
+        }
+        float ao[6][4];
+        float light[6][4];
+        occlusion(neighbors, lights, shades, ao, light);
+        if (is_plant(ew)) {
+            total = 4;
+            float min_ao = 1;
+            float max_light = 0;
+            for (int a = 0; a < 6; a++) {
+                for (int b = 0; b < 4; b++) {
+                    min_ao = MIN(min_ao, ao[a][b]);
+                    max_light = MAX(max_light, light[a][b]);
+                }
+            }
+            float rotation = simplex2(ex, ez, 4, 0.5, 2) * 360;
+            make_plant(
+                data + offset, min_ao, max_light,
+                ex, ey, ez, 0.5, ew, rotation);
+        }
+        else {
+            make_cube(
+                data + offset, ao, light,
+                f1, f2, f3, f4, f5, f6,
+                ex, ey, ez, 0.5, ew);
+        }
+        offset += total * 60;
+    } END_MAP_FOR_EACH;
+
+    free(opaque);
+    free(light);
+    free(highest);
+
+    item->miny = miny;
+    item->maxy = maxy;
+    item->faces = faces;
+    item->data = data;
+} // compute_chunk
+
+void craft::generate_chunk(Chunk *chunk, WorkerItem *item) {
+    chunk->miny = item->miny;
+    chunk->maxy = item->maxy;
+    chunk->faces = item->faces;
+    del_buffer(chunk->buffer);
+    chunk->buffer = gen_faces(10, item->faces, item->data);
+    gen_sign_buffer(chunk);
+}
+
+void craft::gen_chunk_buffer(Chunk *chunk) {
+    WorkerItem _item;
+    WorkerItem *item = &_item;
+    item->p = chunk->p;
+    item->q = chunk->q;
+    for (int dp = -1; dp <= 1; dp++) {
+        for (int dq = -1; dq <= 1; dq++) {
+            Chunk *other = chunk;
+            if (dp || dq) {
+                other = find_chunk(chunk->p + dp, chunk->q + dq);
+            }
+            if (other) {
+                item->block_maps[dp + 1][dq + 1] = &other->map;
+                item->light_maps[dp + 1][dq + 1] = &other->lights;
+            }
+            else {
+                item->block_maps[dp + 1][dq + 1] = 0;
+                item->light_maps[dp + 1][dq + 1] = 0;
+            }
+        }
+    }
+    compute_chunk(item);
+    generate_chunk(chunk, item);
+    chunk->dirty = 0;
+}
+
+void craft::map_set_func(int x, int y, int z, int w, void *arg) {
+    Map *map = (Map *)arg;
+    map_set(map, x, y, z, w);
+}
+
+void craft::load_chunk(WorkerItem *item) {
+    int p = item->p;
+    int q = item->q;
+    Map *block_map = item->block_maps[1][1];
+    Map *light_map = item->light_maps[1][1];
+    create_world(p, q, (world_func) &map_set_func, block_map);
+    db_load_blocks(block_map, p, q);
+    db_load_lights(light_map, p, q);
+}
+
+void craft::request_chunk(int p, int q) {
+    int key = db_get_key(p, q);
+    client_chunk(p, q, key);
+}
+
+void craft::init_chunk(Chunk *chunk, int p, int q) {
+    chunk->p = p;
+    chunk->q = q;
+    chunk->faces = 0;
+    chunk->sign_faces = 0;
+    chunk->buffer = 0;
+    chunk->sign_buffer = 0;
+    dirty_chunk(chunk);
+    SignList *signs = &chunk->signs;
+    sign_list_alloc(signs, 16);
+    db_load_signs(signs, p, q);
+    Map *block_map = &chunk->map;
+    Map *light_map = &chunk->lights;
+    int dx = p * CHUNK_SIZE - 1;
+    int dy = 0;
+    int dz = q * CHUNK_SIZE - 1;
+    map_alloc(block_map, dx, dy, dz, 0x7fff);
+    map_alloc(light_map, dx, dy, dz, 0xf);
+}
+
+void craft::create_chunk(Chunk *chunk, int p, int q) {
+    init_chunk(chunk, p, q);
+
+    WorkerItem _item;
+    WorkerItem *item = &_item;
+    item->p = chunk->p;
+    item->q = chunk->q;
+    item->block_maps[1][1] = &chunk->map;
+    item->light_maps[1][1] = &chunk->lights;
+    load_chunk(item);
+
+    request_chunk(p, q);
+}
+
+void craft::delete_chunks() {
+    int count = g->chunk_count;
+    State *s1 = &g->players->state;
+    State *s2 = &(g->players + g->observe1)->state;
+    State *s3 = &(g->players + g->observe2)->state;
+    State *states[3] = {s1, s2, s3};
+    for (int i = 0; i < count; i++) {
+        Chunk *chunk = g->chunks + i;
+        int remove_chunk = 1;
+        for (int j = 0; j < 3; j++) {
+            State *s = states[j];
+            int p = chunked(s->x);
+            int q = chunked(s->z);
+            if (chunk_distance(chunk, p, q) < g->delete_radius) {
+                remove_chunk = 0;
+                break;
+            }
+        }
+        if (remove_chunk) {
+            map_free(&chunk->map);
+            map_free(&chunk->lights);
+            sign_list_free(&chunk->signs);
+            del_buffer(chunk->buffer);
+            del_buffer(chunk->sign_buffer);
+            Chunk *other = g->chunks + (--count);
+            memcpy(chunk, other, sizeof(Chunk));
+        }
+    }
+    g->chunk_count = count;
+}
+
+void craft::delete_all_chunks() {
+    for (int i = 0; i < g->chunk_count; i++) {
+        Chunk *chunk = g->chunks + i;
+        map_free(&chunk->map);
+        map_free(&chunk->lights);
+        sign_list_free(&chunk->signs);
+        del_buffer(chunk->buffer);
+        del_buffer(chunk->sign_buffer);
+    }
+    g->chunk_count = 0;
+}
+
+void craft::check_workers() {
+    for (int i = 0; i < WORKERS; i++) {
+        Worker *worker = g->workers + i;
+        mtx_lock(&worker->mtx);
+        if (worker->state == WORKER_DONE) {
+            WorkerItem *item = &worker->item;
+            Chunk *chunk = find_chunk(item->p, item->q);
+            if (chunk) {
+                if (item->load) {
+                    Map *block_map = item->block_maps[1][1];
+                    Map *light_map = item->light_maps[1][1];
+                    map_free(&chunk->map);
+                    map_free(&chunk->lights);
+                    map_copy(&chunk->map, block_map);
+                    map_copy(&chunk->lights, light_map);
+                    request_chunk(item->p, item->q);
+                }
+                generate_chunk(chunk, item);
+            }
+            for (int a = 0; a < 3; a++) {
+                for (int b = 0; b < 3; b++) {
+                    Map *block_map = item->block_maps[a][b];
+                    Map *light_map = item->light_maps[a][b];
+                    if (block_map) {
+                        map_free(block_map);
+                        free(block_map);
+                    }
+                    if (light_map) {
+                        map_free(light_map);
+                        free(light_map);
+                    }
+                }
+            }
+            worker->state = WORKER_IDLE;
+        }
+        mtx_unlock(&worker->mtx);
+    }
+}
+
+void craft::force_chunks(Player *player) {
+    State *s = &player->state;
+    int p = chunked(s->x);
+    int q = chunked(s->z);
+    int r = 1;
+    for (int dp = -r; dp <= r; dp++) {
+        for (int dq = -r; dq <= r; dq++) {
+            int a = p + dp;
+            int b = q + dq;
+            Chunk *chunk = find_chunk(a, b);
+            if (chunk) {
+                if (chunk->dirty) {
+                    gen_chunk_buffer(chunk);
+                }
+            }
+            else if (g->chunk_count < MAX_CHUNKS) {
+                chunk = g->chunks + g->chunk_count++;
+                create_chunk(chunk, a, b);
+                gen_chunk_buffer(chunk);
+            }
+        }
+    }
+}
+
+void craft::ensure_chunks_worker(Player *player, Worker *worker) {
+    State *s = &player->state;
+    float matrix[16];
+    set_matrix_3d(
+        matrix, g->width, g->height,
+        s->x, s->y, s->z, s->rx, s->ry, g->fov, g->ortho, g->render_radius);
+    float planes[6][4];
+    frustum_planes(planes, g->render_radius, matrix);
+    int p = chunked(s->x);
+    int q = chunked(s->z);
+    int r = g->create_radius;
+    int start = 0x0fffffff;
+    int best_score = start;
+    int best_a = 0;
+    int best_b = 0;
+    for (int dp = -r; dp <= r; dp++) {
+        for (int dq = -r; dq <= r; dq++) {
+            int a = p + dp;
+            int b = q + dq;
+            int index = (ABS(a) ^ ABS(b)) % WORKERS;
+            if (index != worker->index) {
+                continue;
+            }
+            Chunk *chunk = find_chunk(a, b);
+            if (chunk && !chunk->dirty) {
+                continue;
+            }
+            int distance = MAX(ABS(dp), ABS(dq));
+            int invisible = !chunk_visible(planes, a, b, 0, 256);
+            int priority = 0;
+            if (chunk) {
+                priority = chunk->buffer && chunk->dirty;
+            }
+            int score = (invisible << 24) | (priority << 16) | distance;
+            if (score < best_score) {
+                best_score = score;
+                best_a = a;
+                best_b = b;
+            }
+        }
+    }
+    if (best_score == start) {
+        return;
+    }
+    int a = best_a;
+    int b = best_b;
+    int load = 0;
+    Chunk *chunk = find_chunk(a, b);
+    if (!chunk) {
+        load = 1;
+        if (g->chunk_count < MAX_CHUNKS) {
+            chunk = g->chunks + g->chunk_count++;
+            init_chunk(chunk, a, b);
+        }
+        else {
+            return;
+        }
+    }
+    WorkerItem *item = &worker->item;
+    item->p = chunk->p;
+    item->q = chunk->q;
+    item->load = load;
+    for (int dp = -1; dp <= 1; dp++) {
+        for (int dq = -1; dq <= 1; dq++) {
+            Chunk *other = chunk;
+            if (dp || dq) {
+                other = find_chunk(chunk->p + dp, chunk->q + dq);
+            }
+            if (other) {
+                Map *block_map = (Map*) malloc(sizeof(Map));
+                map_copy(block_map, &other->map);
+                Map *light_map = (Map*) malloc(sizeof(Map));
+                map_copy(light_map, &other->lights);
+                item->block_maps[dp + 1][dq + 1] = block_map;
+                item->light_maps[dp + 1][dq + 1] = light_map;
+            }
+            else {
+                item->block_maps[dp + 1][dq + 1] = 0;
+                item->light_maps[dp + 1][dq + 1] = 0;
+            }
+        }
+    }
+    chunk->dirty = 0;
+    worker->state = WORKER_BUSY;
+    cnd_signal(&worker->cnd);
+}
+
+void craft::ensure_chunks(Player *player) {
+    check_workers();
+    force_chunks(player);
+    for (int i = 0; i < WORKERS; i++) {
+        Worker *worker = g->workers + i;
+        mtx_lock(&worker->mtx);
+        if (worker->state == WORKER_IDLE) {
+            ensure_chunks_worker(player, worker);
+        }
+        mtx_unlock(&worker->mtx);
+    }
+}
+
+int craft::worker_run(void *arg) {
+    Worker *worker = (Worker *)arg;
+    int running = 1;
+    while (running) {
+        mtx_lock(&worker->mtx);
+        while (worker->state != WORKER_BUSY) {
+            cnd_wait(&worker->cnd, &worker->mtx);
+        }
+        mtx_unlock(&worker->mtx);
+        WorkerItem *item = &worker->item;
+        if (item->load) {
+            load_chunk(item);
+        }
+        compute_chunk(item);
+        mtx_lock(&worker->mtx);
+        worker->state = WORKER_DONE;
+        mtx_unlock(&worker->mtx);
+    }
+    return 0;
+}
+
+void craft::unset_sign(int x, int y, int z) {
+    int p = chunked(x);
+    int q = chunked(z);
+    Chunk *chunk = find_chunk(p, q);
+    if (chunk) {
+        SignList *signs = &chunk->signs;
+        if (sign_list_remove_all(signs, x, y, z)) {
+            chunk->dirty = 1;
+            db_delete_signs(x, y, z);
+        }
+    }
+    else {
+        db_delete_signs(x, y, z);
+    }
+}
+
+void craft::unset_sign_face(int x, int y, int z, int face) {
+    int p = chunked(x);
+    int q = chunked(z);
+    Chunk *chunk = find_chunk(p, q);
+    if (chunk) {
+        SignList *signs = &chunk->signs;
+        if (sign_list_remove(signs, x, y, z, face)) {
+            chunk->dirty = 1;
+            db_delete_sign(x, y, z, face);
+        }
+    }
+    else {
+        db_delete_sign(x, y, z, face);
+    }
+}
+
+void craft::_set_sign(int p, int q, int x, int y, int z, int face, const char *text, int dirty) {
+    if (strlen(text) == 0) {
+        unset_sign_face(x, y, z, face);
+        return;
+    }
+    Chunk *chunk = find_chunk(p, q);
+    if (chunk) {
+        SignList *signs = &chunk->signs;
+        sign_list_add(signs, x, y, z, face, text);
+        if (dirty) {
+            chunk->dirty = 1;
+        }
+    }
+    db_insert_sign(p, q, x, y, z, face, text);
+}
+
+void craft::set_sign(int x, int y, int z, int face, const char *text) {
+    int p = chunked(x);
+    int q = chunked(z);
+    _set_sign(p, q, x, y, z, face, text, 1);
+    client_sign(x, y, z, face, text);
+}
+
+void craft::toggle_light(int x, int y, int z) {
+    int p = chunked(x);
+    int q = chunked(z);
+    Chunk *chunk = find_chunk(p, q);
+    if (chunk) {
+        Map *map = &chunk->lights;
+        int w = map_get(map, x, y, z) ? 0 : 15;
+        map_set(map, x, y, z, w);
+        db_insert_light(p, q, x, y, z, w);
+        client_light(x, y, z, w);
+        dirty_chunk(chunk);
+    }
+}
+
+void craft::set_light(int p, int q, int x, int y, int z, int w) {
+    Chunk *chunk = find_chunk(p, q);
+    if (chunk) {
+        Map *map = &chunk->lights;
+        if (map_set(map, x, y, z, w)) {
+            dirty_chunk(chunk);
+            db_insert_light(p, q, x, y, z, w);
+        }
+    }
+    else {
+        db_insert_light(p, q, x, y, z, w);
+    }
+}
+
+void craft::_set_block(int p, int q, int x, int y, int z, int w, int dirty) {
+    Chunk *chunk = find_chunk(p, q);
+    if (chunk) {
+        Map *map = &chunk->map;
+        if (map_set(map, x, y, z, w)) {
+            if (dirty) {
+                dirty_chunk(chunk);
+            }
+            db_insert_block(p, q, x, y, z, w);
+        }
+    }
+    else {
+        db_insert_block(p, q, x, y, z, w);
+    }
+    if (w == 0 && chunked(x) == p && chunked(z) == q) {
+        unset_sign(x, y, z);
+        set_light(p, q, x, y, z, 0);
+    }
+}
+
+void craft::set_block(int x, int y, int z, int w) {
+    int p = chunked(x);
+    int q = chunked(z);
+    _set_block(p, q, x, y, z, w, 1);
+    for (int dx = -1; dx <= 1; dx++) {
+        for (int dz = -1; dz <= 1; dz++) {
+            if (dx == 0 && dz == 0) {
+                continue;
+            }
+            if (dx && chunked(x + dx) == p) {
+                continue;
+            }
+            if (dz && chunked(z + dz) == q) {
+                continue;
+            }
+            _set_block(p + dx, q + dz, x, y, z, -w, 1);
+        }
+    }
+    client_block(x, y, z, w);
+}
+
+void craft::record_block(int x, int y, int z, int w) {
+    memcpy(&g->block1, &g->block0, sizeof(Block));
+    g->block0.x = x;
+    g->block0.y = y;
+    g->block0.z = z;
+    g->block0.w = w;
+}
+
+int craft::get_block(int x, int y, int z) {
+    int p = chunked(x);
+    int q = chunked(z);
+    Chunk *chunk = find_chunk(p, q);
+    if (chunk) {
+        Map *map = &chunk->map;
+        return map_get(map, x, y, z);
+    }
+    return 0;
+}
+
+void craft::builder_block(int x, int y, int z, int w) {
+    if (y <= 0 || y >= 256) {
+        return;
+    }
+    if (is_destructable(get_block(x, y, z))) {
+        set_block(x, y, z, 0);
+    }
+    if (w) {
+        set_block(x, y, z, w);
+    }
+}
+
+int craft::render_chunks(Attrib *attrib, Player *player) {
+    int result = 0;
+    State *s = &player->state;
+    ensure_chunks(player);
+    int p = chunked(s->x);
+    int q = chunked(s->z);
+    float light = get_daylight();
+    float matrix[16];
+    set_matrix_3d(
+        matrix, g->width, g->height,
+        s->x, s->y, s->z, s->rx, s->ry, g->fov, g->ortho, g->render_radius);
+    float planes[6][4];
+    frustum_planes(planes, g->render_radius, matrix);
+    glUseProgram(attrib->program);
+    glUniformMatrix4fv(attrib->matrix, 1, GL_FALSE, matrix);
+    glUniform3f(attrib->camera, s->x, s->y, s->z);
+    glUniform1i(attrib->sampler, 0);
+    glUniform1i(attrib->extra1, 2);
+    glUniform1f(attrib->extra2, light);
+    glUniform1f(attrib->extra3, g->render_radius * CHUNK_SIZE);
+    glUniform1i(attrib->extra4, g->ortho);
+    glUniform1f(attrib->timer, time_of_day());
+    for (int i = 0; i < g->chunk_count; i++) {
+        Chunk *chunk = g->chunks + i;
+        if (chunk_distance(chunk, p, q) > g->render_radius) {
+            continue;
+        }
+        if (!chunk_visible(
+            planes, chunk->p, chunk->q, chunk->miny, chunk->maxy))
+        {
+            continue;
+        }
+        draw_chunk(attrib, chunk);
+        result += chunk->faces;
+    }
+    return result;
+}
+
+void craft::render_signs(Attrib *attrib, Player *player) {
+    State *s = &player->state;
+    int p = chunked(s->x);
+    int q = chunked(s->z);
+    float matrix[16];
+    set_matrix_3d(
+        matrix, g->width, g->height,
+        s->x, s->y, s->z, s->rx, s->ry, g->fov, g->ortho, g->render_radius);
+    float planes[6][4];
+    frustum_planes(planes, g->render_radius, matrix);
+    glUseProgram(attrib->program);
+    glUniformMatrix4fv(attrib->matrix, 1, GL_FALSE, matrix);
+    glUniform1i(attrib->sampler, 3);
+    glUniform1i(attrib->extra1, 1);
+    for (int i = 0; i < g->chunk_count; i++) {
+        Chunk *chunk = g->chunks + i;
+        if (chunk_distance(chunk, p, q) > g->sign_radius) {
+            continue;
+        }
+        if (!chunk_visible(
+            planes, chunk->p, chunk->q, chunk->miny, chunk->maxy))
+        {
+            continue;
+        }
+        draw_signs(attrib, chunk);
+    }
+}
+
+void craft::render_sign(Attrib *attrib, Player *player) {
+    if (!g->typing || g->typing_buffer[0] != CRAFT_KEY_SIGN) {
+        return;
+    }
+    int x, y, z, face;
+    if (!hit_test_face(player, &x, &y, &z, &face)) {
+        return;
+    }
+    State *s = &player->state;
+    float matrix[16];
+    set_matrix_3d(
+        matrix, g->width, g->height,
+        s->x, s->y, s->z, s->rx, s->ry, g->fov, g->ortho, g->render_radius);
+    glUseProgram(attrib->program);
+    glUniformMatrix4fv(attrib->matrix, 1, GL_FALSE, matrix);
+    glUniform1i(attrib->sampler, 3);
+    glUniform1i(attrib->extra1, 1);
+    char text[MAX_SIGN_LENGTH];
+    strncpy(text, g->typing_buffer + 1, MAX_SIGN_LENGTH);
+    text[MAX_SIGN_LENGTH - 1] = '\0';
+    GLfloat *data = malloc_faces(5, strlen(text));
+    int length = _gen_sign_buffer(data, x, y, z, face, text);
+    GLuint buffer = gen_faces(5, length, data);
+    draw_sign(attrib, buffer, length);
+    del_buffer(buffer);
+}
+
+void craft::render_players(Attrib *attrib, Player *player) {
+    State *s = &player->state;
+    float matrix[16];
+    set_matrix_3d(
+        matrix, g->width, g->height,
+        s->x, s->y, s->z, s->rx, s->ry, g->fov, g->ortho, g->render_radius);
+    glUseProgram(attrib->program);
+    glUniformMatrix4fv(attrib->matrix, 1, GL_FALSE, matrix);
+    glUniform3f(attrib->camera, s->x, s->y, s->z);
+    glUniform1i(attrib->sampler, 0);
+    glUniform1f(attrib->timer, time_of_day());
+    for (int i = 0; i < g->player_count; i++) {
+        Player *other = g->players + i;
+        if (other != player) {
+            draw_player(attrib, other);
+        }
+    }
+}
+
+void craft::render_sky(Attrib *attrib, Player *player, GLuint buffer) {
+    State *s = &player->state;
+    float matrix[16];
+    set_matrix_3d(
+        matrix, g->width, g->height,
+        0, 0, 0, s->rx, s->ry, g->fov, 0, g->render_radius);
+    
+    glUseProgram(attrib->program);
+    glUniformMatrix4fv(attrib->matrix, 1, GL_FALSE, matrix);
+    glUniform1i(attrib->sampler, 2);
+    glUniform1f(attrib->timer, time_of_day());
+    
+    draw_triangles_3d(attrib, buffer, 512 * 3);
+}
+
+void craft::render_wireframe(Attrib *attrib, Player *player) {
+    State *s = &player->state;
+    float matrix[16];
+    set_matrix_3d(
+        matrix, g->width, g->height,
+        s->x, s->y, s->z, s->rx, s->ry, g->fov, g->ortho, g->render_radius);
+    int hx, hy, hz;
+    int hw = hit_test(0, s->x, s->y, s->z, s->rx, s->ry, &hx, &hy, &hz);
+    if (is_obstacle(hw)) {
+        glUseProgram(attrib->program);
+        glLineWidth(1);
+        glEnable(GL_COLOR_LOGIC_OP);
+        glUniformMatrix4fv(attrib->matrix, 1, GL_FALSE, matrix);
+        GLuint wireframe_buffer = gen_wireframe_buffer(hx, hy, hz, 0.53);
+        draw_lines(attrib, wireframe_buffer, 3, 24);
+        del_buffer(wireframe_buffer);
+        glDisable(GL_COLOR_LOGIC_OP);
+    }
+}
+
+void craft::render_crosshairs(Attrib *attrib) {
+    float matrix[16];
+    set_matrix_2d(matrix, g->width, g->height);
+    glUseProgram(attrib->program);
+    glLineWidth(4 * g->scale);
+    glEnable(GL_COLOR_LOGIC_OP);
+    glUniformMatrix4fv(attrib->matrix, 1, GL_FALSE, matrix);
+    GLuint crosshair_buffer = gen_crosshair_buffer();
+    draw_lines(attrib, crosshair_buffer, 2, 4);
+    del_buffer(crosshair_buffer);
+    glDisable(GL_COLOR_LOGIC_OP);
+}
+
+void craft::render_item(Attrib *attrib) {
+    float matrix[16];
+    set_matrix_item(matrix, g->width, g->height, g->scale);
+    glUseProgram(attrib->program);
+    glUniformMatrix4fv(attrib->matrix, 1, GL_FALSE, matrix);
+    glUniform3f(attrib->camera, 0, 0, 5);
+    glUniform1i(attrib->sampler, 0);
+    glUniform1f(attrib->timer, time_of_day());
+    int w = items[g->item_index];
+    if (is_plant(w)) {
+        GLuint buffer = gen_plant_buffer(0, 0, 0, 0.5, w);
+        draw_plant(attrib, buffer);
+        del_buffer(buffer);
+    }
+    else {
+        GLuint buffer = gen_cube_buffer(0, 0, 0, 0.5, w);
+        draw_cube(attrib, buffer);
+        del_buffer(buffer);
+    }
+}
+
+void craft::render_text(Attrib *attrib, int justify, float x, float y, float n, char *text) {
+    float matrix[16];
+    set_matrix_2d(matrix, g->width, g->height);
+    glUseProgram(attrib->program);
+    glUniformMatrix4fv(attrib->matrix, 1, GL_FALSE, matrix);
+    glUniform1i(attrib->sampler, 1);
+    glUniform1i(attrib->extra1, 0);
+    int length = strlen(text);
+    x -= n * justify * (length - 1) / 2;
+    GLuint buffer = gen_text_buffer(x, y, n, text);
+    draw_text(attrib, buffer, length);
+    del_buffer(buffer);
+}
+
+void craft::add_message(const char *text) {
+    printf("%s\n", text);
+    snprintf(
+        g->messages[g->message_index], MAX_TEXT_LENGTH, "%s", text);
+    g->message_index = (g->message_index + 1) % MAX_MESSAGES;
+}
+
+void craft::login() {
+    char username[128] = {0};
+    char identity_token[128] = {0};
+    char access_token[128] = {0};
+    if (db_auth_get_selected(username, 128, identity_token, 128)) {
+        printf("Contacting login server for username: %s\n", username);
+        if (get_access_token(
+            access_token, 128, username, identity_token))
+        {
+            printf("Successfully authenticated with the login server\n");
+            client_login(username, access_token);
+        }
+        else {
+            printf("Failed to authenticate with the login server\n");
+            client_login("", "");
+        }
+    }
+    else {
+        printf("Logging in anonymously\n");
+        client_login("", "");
+    }
+}
+
+void craft::copy() {
+    memcpy(&g->copy0, &g->block0, sizeof(Block));
+    memcpy(&g->copy1, &g->block1, sizeof(Block));
+}
+
+void craft::paste() {
+    Block *c1 = &g->copy1;
+    Block *c2 = &g->copy0;
+    Block *p1 = &g->block1;
+    Block *p2 = &g->block0;
+    int scx = SIGN(c2->x - c1->x);
+    int scz = SIGN(c2->z - c1->z);
+    int spx = SIGN(p2->x - p1->x);
+    int spz = SIGN(p2->z - p1->z);
+    int oy = p1->y - c1->y;
+    int dx = ABS(c2->x - c1->x);
+    int dz = ABS(c2->z - c1->z);
+    for (int y = 0; y < 256; y++) {
+        for (int x = 0; x <= dx; x++) {
+            for (int z = 0; z <= dz; z++) {
+                int w = get_block(c1->x + x * scx, y, c1->z + z * scz);
+                builder_block(p1->x + x * spx, y + oy, p1->z + z * spz, w);
+            }
+        }
+    }
+}
+
+void craft::array(Block *b1, Block *b2, int xc, int yc, int zc) {
+    if (b1->w != b2->w) {
+        return;
+    }
+    int w = b1->w;
+    int dx = b2->x - b1->x;
+    int dy = b2->y - b1->y;
+    int dz = b2->z - b1->z;
+    xc = dx ? xc : 1;
+    yc = dy ? yc : 1;
+    zc = dz ? zc : 1;
+    for (int i = 0; i < xc; i++) {
+        int x = b1->x + dx * i;
+        for (int j = 0; j < yc; j++) {
+            int y = b1->y + dy * j;
+            for (int k = 0; k < zc; k++) {
+                int z = b1->z + dz * k;
+                builder_block(x, y, z, w);
+            }
+        }
+    }
+}
+
+void craft::cube(Block *b1, Block *b2, int fill) {
+    if (b1->w != b2->w) {
+        return;
+    }
+    int w = b1->w;
+    int x1 = MIN(b1->x, b2->x);
+    int y1 = MIN(b1->y, b2->y);
+    int z1 = MIN(b1->z, b2->z);
+    int x2 = MAX(b1->x, b2->x);
+    int y2 = MAX(b1->y, b2->y);
+    int z2 = MAX(b1->z, b2->z);
+    int a = (x1 == x2) + (y1 == y2) + (z1 == z2);
+    for (int x = x1; x <= x2; x++) {
+        for (int y = y1; y <= y2; y++) {
+            for (int z = z1; z <= z2; z++) {
+                if (!fill) {
+                    int n = 0;
+                    n += x == x1 || x == x2;
+                    n += y == y1 || y == y2;
+                    n += z == z1 || z == z2;
+                    if (n <= a) {
+                        continue;
+                    }
+                }
+                builder_block(x, y, z, w);
+            }
+        }
+    }
+}
+
+void craft::sphere(Block *center, int radius, int fill, int fx, int fy, int fz) {
+    static const float offsets[8][3] = {
+        {-0.5, -0.5, -0.5},
+        {-0.5, -0.5, 0.5},
+        {-0.5, 0.5, -0.5},
+        {-0.5, 0.5, 0.5},
+        {0.5, -0.5, -0.5},
+        {0.5, -0.5, 0.5},
+        {0.5, 0.5, -0.5},
+        {0.5, 0.5, 0.5}
+    };
+    int cx = center->x;
+    int cy = center->y;
+    int cz = center->z;
+    int w = center->w;
+    for (int x = cx - radius; x <= cx + radius; x++) {
+        if (fx && x != cx) {
+            continue;
+        }
+        for (int y = cy - radius; y <= cy + radius; y++) {
+            if (fy && y != cy) {
+                continue;
+            }
+            for (int z = cz - radius; z <= cz + radius; z++) {
+                if (fz && z != cz) {
+                    continue;
+                }
+                int inside = 0;
+                int outside = fill;
+                for (int i = 0; i < 8; i++) {
+                    float dx = x + offsets[i][0] - cx;
+                    float dy = y + offsets[i][1] - cy;
+                    float dz = z + offsets[i][2] - cz;
+                    float d = SDL_sqrt(dx * dx + dy * dy + dz * dz);
+                    if (d < radius) {
+                        inside = 1;
+                    }
+                    else {
+                        outside = 1;
+                    }
+                }
+                if (inside && outside) {
+                    builder_block(x, y, z, w);
+                }
+            }
+        }
+    }
+}
+
+void craft::cylinder(Block *b1, Block *b2, int radius, int fill) {
+    if (b1->w != b2->w) {
+        return;
+    }
+    int w = b1->w;
+    int x1 = MIN(b1->x, b2->x);
+    int y1 = MIN(b1->y, b2->y);
+    int z1 = MIN(b1->z, b2->z);
+    int x2 = MAX(b1->x, b2->x);
+    int y2 = MAX(b1->y, b2->y);
+    int z2 = MAX(b1->z, b2->z);
+    int fx = x1 != x2;
+    int fy = y1 != y2;
+    int fz = z1 != z2;
+    if (fx + fy + fz != 1) {
+        return;
+    }
+    Block block = {x1, y1, z1, w};
+    if (fx) {
+        for (int x = x1; x <= x2; x++) {
+            block.x = x;
+            sphere(&block, radius, fill, 1, 0, 0);
+        }
+    }
+    if (fy) {
+        for (int y = y1; y <= y2; y++) {
+            block.y = y;
+            sphere(&block, radius, fill, 0, 1, 0);
+        }
+    }
+    if (fz) {
+        for (int z = z1; z <= z2; z++) {
+            block.z = z;
+            sphere(&block, radius, fill, 0, 0, 1);
+        }
+    }
+}
+
+void craft::tree(Block *block) {
+    int bx = block->x;
+    int by = block->y;
+    int bz = block->z;
+    for (int y = by + 3; y < by + 8; y++) {
+        for (int dx = -3; dx <= 3; dx++) {
+            for (int dz = -3; dz <= 3; dz++) {
+                int dy = y - (by + 4);
+                int d = (dx * dx) + (dy * dy) + (dz * dz);
+                if (d < 11) {
+                    builder_block(bx + dx, y, bz + dz, 15);
+                }
+            }
+        }
+    }
+    for (int y = by; y < by + 7; y++) {
+        builder_block(bx, y, bz, 5);
+    }
+}
+
+void craft::parse_command(const char *buffer, int forward) {
+    char username[128] = {0};
+    char token[128] = {0};
+    char server_addr[MAX_ADDR_LENGTH];
+    int server_port = DEFAULT_PORT;
+    char filename[MAX_PATH_LENGTH];
+    int radius, count, xc, yc, zc;
+    if (sscanf(buffer, "/identity %128s %128s", username, token) == 2) {
+        db_auth_set(username, token);
+        add_message("Successfully imported identity token!");
+        login();
+    }
+    else if (strcmp(buffer, "/logout") == 0) {
+        db_auth_select_none();
+        login();
+    }
+    else if (sscanf(buffer, "/login %128s", username) == 1) {
+        if (db_auth_select(username)) {
+            login();
+        }
+        else {
+            add_message("Unknown username.");
+        }
+    }
+    else if (sscanf(buffer,
+        "/online %128s %d", server_addr, &server_port) >= 1)
+    {
+        g->mode_changed = 1;
+        g->mode = MODE_ONLINE;
+        strncpy(g->server_addr, server_addr, MAX_ADDR_LENGTH);
+        g->server_port = server_port;
+        snprintf(g->db_path, MAX_PATH_LENGTH,
+            "cache.%s.%d.db", g->server_addr, g->server_port);
+    }
+    else if (sscanf(buffer, "/offline %128s", filename) == 1) {
+        g->mode_changed = 1;
+        g->mode = MODE_OFFLINE;
+        snprintf(g->db_path, MAX_PATH_LENGTH, "%s.db", filename);
+    }
+    else if (strcmp(buffer, "/offline") == 0) {
+        g->mode_changed = 1;
+        g->mode = MODE_OFFLINE;
+        snprintf(g->db_path, MAX_PATH_LENGTH, "%s", DB_PATH);
+    }
+    else if (sscanf(buffer, "/view %d", &radius) == 1) {
+        if (radius >= 1 && radius <= 24) {
+            g->create_radius = radius;
+            g->render_radius = radius;
+            g->delete_radius = radius + 4;
+        }
+        else {
+            add_message("Viewing distance must be between 1 and 24.");
+        }
+    }
+    else if (strcmp(buffer, "/copy") == 0) {
+        copy();
+    }
+    else if (strcmp(buffer, "/paste") == 0) {
+        paste();
+    }
+    else if (strcmp(buffer, "/tree") == 0) {
+        tree(&g->block0);
+    }
+    else if (sscanf(buffer, "/array %d %d %d", &xc, &yc, &zc) == 3) {
+        array(&g->block1, &g->block0, xc, yc, zc);
+    }
+    else if (sscanf(buffer, "/array %d", &count) == 1) {
+        array(&g->block1, &g->block0, count, count, count);
+    }
+    else if (strcmp(buffer, "/fcube") == 0) {
+        cube(&g->block0, &g->block1, 1);
+    }
+    else if (strcmp(buffer, "/cube") == 0) {
+        cube(&g->block0, &g->block1, 0);
+    }
+    else if (sscanf(buffer, "/fsphere %d", &radius) == 1) {
+        sphere(&g->block0, radius, 1, 0, 0, 0);
+    }
+    else if (sscanf(buffer, "/sphere %d", &radius) == 1) {
+        sphere(&g->block0, radius, 0, 0, 0, 0);
+    }
+    else if (sscanf(buffer, "/fcirclex %d", &radius) == 1) {
+        sphere(&g->block0, radius, 1, 1, 0, 0);
+    }
+    else if (sscanf(buffer, "/circlex %d", &radius) == 1) {
+        sphere(&g->block0, radius, 0, 1, 0, 0);
+    }
+    else if (sscanf(buffer, "/fcircley %d", &radius) == 1) {
+        sphere(&g->block0, radius, 1, 0, 1, 0);
+    }
+    else if (sscanf(buffer, "/circley %d", &radius) == 1) {
+        sphere(&g->block0, radius, 0, 0, 1, 0);
+    }
+    else if (sscanf(buffer, "/fcirclez %d", &radius) == 1) {
+        sphere(&g->block0, radius, 1, 0, 0, 1);
+    }
+    else if (sscanf(buffer, "/circlez %d", &radius) == 1) {
+        sphere(&g->block0, radius, 0, 0, 0, 1);
+    }
+    else if (sscanf(buffer, "/fcylinder %d", &radius) == 1) {
+        cylinder(&g->block0, &g->block1, radius, 1);
+    }
+    else if (sscanf(buffer, "/cylinder %d", &radius) == 1) {
+        cylinder(&g->block0, &g->block1, radius, 0);
+    }
+    else if (forward) {
+        client_talk(buffer);
+    }
+}
+
+void craft::on_light() {
+    State *s = &g->players->state;
+    int hx, hy, hz;
+    int hw = hit_test(0, s->x, s->y, s->z, s->rx, s->ry, &hx, &hy, &hz);
+    if (hy > 0 && hy < 256 && is_destructable(hw)) {
+        toggle_light(hx, hy, hz);
+    }
+}
+
+void craft::on_left_click() {
+    State *s = &g->players->state;
+    int hx, hy, hz;
+    int hw = hit_test(0, s->x, s->y, s->z, s->rx, s->ry, &hx, &hy, &hz);
+    if (hy > 0 && hy < 256 && is_destructable(hw)) {
+        set_block(hx, hy, hz, 0);
+        record_block(hx, hy, hz, 0);
+        if (is_plant(get_block(hx, hy + 1, hz))) {
+            set_block(hx, hy + 1, hz, 0);
+        }
+    }
+}
+
+void craft::on_right_click() {
+    State *s = &g->players->state;
+    int hx, hy, hz;
+    int hw = hit_test(1, s->x, s->y, s->z, s->rx, s->ry, &hx, &hy, &hz);
+    if (hy > 0 && hy < 256 && is_obstacle(hw)) {
+        if (!player_intersects_block(2, s->x, s->y, s->z, hx, hy, hz)) {
+            set_block(hx, hy, hz, items[g->item_index]);
+            record_block(hx, hy, hz, items[g->item_index]);
+        }
+    }
+}
+
+void craft::on_middle_click() {
+    State *s = &g->players->state;
+    int hx, hy, hz;
+    int hw = hit_test(0, s->x, s->y, s->z, s->rx, s->ry, &hx, &hy, &hz);
+    for (int i = 0; i < item_count; i++) {
+        if (items[i] == hw) {
+            g->item_index = i;
+            break;
+        }
+    }
+}
+
+// https://github.com/rswinkle/Craft/blob/sdl/src/main.c
+int craft::handle_events(double dt) {
+    // Handle window init and force a scale
+    static SDL_bool first_event = 1;
+	static float dy = 0;
+	State* s = &g->players->state;
+	int sz = 0;
+	int sx = 0;
+	float m = 0.0025;
+
+	SDL_Event e;
+	int sc, code;
+
+	// eat all escapes this frame after copy dialog ended with "no"
+	int copy_escape = 0;
+
+	SDL_Keymod mod_state = SDL_GetModState();
+
+	int control = mod_state ;//& (KMOD_LCTRL | KMOD_RCTRL | KMOD_LGUI | KMOD_RGUI);
+    int exclusive = SDL_GetRelativeMouseMode();
+
+	while (SDL_PollEvent(&e)) {
+		/*
+		if (e.type == g->userevent) {
+
+			code = e.user.code;
+			switch (code) {
+			}
+		}
+		*/
+		switch (e.type) {
+		case SDL_EVENT_QUIT:
+			return 1;
+		case SDL_EVENT_KEY_UP:
+			sc = e.key.keysym.scancode;
+			switch (sc) {
+			case SDL_SCANCODE_ESCAPE:
+				if (g->typing) {
+					g->typing = 0;
+				} else if (exclusive) {
+					SDL_SetRelativeMouseMode(SDL_FALSE);
+				} else {
+					return 1;
+				}
+				break;
+			}
+			break;
+
+		case SDL_EVENT_KEY_DOWN:
+			sc = e.key.keysym.scancode;
+			switch (sc) {
+			case SDL_SCANCODE_RETURN:
+				if (g->typing) {
+					if (mod_state /*& (KMOD_LSHIFT | KMOD_RSHIFT)*/) {
+						if (g->text_len < MAX_TEXT_LENGTH - 1) {
+							g->typing_buffer[g->text_len] = '\r'; // TODO? \n?
+							g->typing_buffer[g->text_len+1] = '\0';
+						}
+					} else {
+						g->typing = 0;
+						if (g->typing_buffer[0] == CRAFT_KEY_SIGN) {
+							Player *player = g->players;
+							int x, y, z, face;
+							if (hit_test_face(player, &x, &y, &z, &face)) {
+								set_sign(x, y, z, face, g->typing_buffer + 1);
+							}
+						} else if (g->typing_buffer[0] == '/') {
+							parse_command(g->typing_buffer, 1);
+						} else {
+							client_talk(g->typing_buffer);
+						}
+					}
+				} else {
+					if (control) {
+						on_right_click();
+					} else {
+						on_left_click();
+					}
+				}
+				break;
+
+			case SDL_SCANCODE_V:
+				if (control) {
+					char* clip_buffer = SDL_GetClipboardText();
+					if (g->typing) {
+						g->suppress_char = 1;
+						strncat(g->typing_buffer, clip_buffer,
+							MAX_TEXT_LENGTH - g->text_len - 1);
+					} else {
+						parse_command(clip_buffer, 0);
+					}
+				}
+				break;
+
+			case SDL_SCANCODE_0:
+				if (!g->typing)
+					g->item_index = 9;
+				break;
+			case SDL_SCANCODE_1:
+			case SDL_SCANCODE_2:
+			case SDL_SCANCODE_3:
+			case SDL_SCANCODE_4:
+			case SDL_SCANCODE_5:
+			case SDL_SCANCODE_6:
+			case SDL_SCANCODE_7:
+			case SDL_SCANCODE_8:
+			case SDL_SCANCODE_9:
+				if (!g->typing)
+					g->item_index = (sc - SDL_SCANCODE_1);
+				break;
+
+			case KEY_FLY:
+				if (!g->typing)
+					g->flying = !g->flying;
+				break;
+
+			case KEY_ITEM_NEXT:
+				if (!g->typing)
+					g->item_index = (g->item_index + 1) % item_count;
+				break;
+			case KEY_ITEM_PREV:
+				if (!g->typing) {
+					g->item_index--;
+					if (g->item_index < 0)
+						g->item_index = item_count - 1;
+				}
+				break;
+			case KEY_OBSERVE:
+				if (!g->typing)
+					g->observe1 = (g->observe1 + 1) % g->player_count;
+				break;
+
+			case KEY_OBSERVE_INSET:
+				if (!g->typing)
+					g->observe2 = (g->observe2 + 1) % g->player_count;
+				break;
+
+				/*
+			case KEY_ORTHO:
+				break;
+			case KEY_ZOOM:
+				break;
+			case KEY_FORWARD:
+				break;
+			case KEY_BACKWARD:
+				break;
+			case KEY_LEFT:
+				break;
+			case KEY_RIGHT:
+				break;
+			case KEY_UP:
+				break;
+			case KEY_DOWN:
+				break;
+				*/
+
+			case KEY_CHAT:
+				g->typing = 1;
+				g->typing_buffer[0] = '\0';
+				g->text_len = 0;
+				SDL_StartTextInput();
+				break;
+
+			case KEY_COMMAND:
+				g->typing = 1;
+				g->typing_buffer[0] = '\0';
+				SDL_StartTextInput();
+				break;
+
+			case KEY_SIGN:
+				g->typing = 1;
+				g->typing_buffer[0] = '\0';
+				SDL_StartTextInput();
+				break;
+
+			}
+
+			break;
+
+		case SDL_EVENT_TEXT_INPUT:
+			// could probably just do text[text_len++] = e.text.text[0]
+			// since I only handle ascii
+			if (g->typing && g->text_len < MAX_TEXT_LENGTH -1) {
+				strcat(g->typing_buffer, e.text.text);
+				g->text_len += strlen(e.text.text);
+				//SDL_Log("text is \"%s\" \"%s\" %d %d\n", g->typing_buffer, composition, cursor, selection_len);
+				//SDL_LogDebug(SDL_LOG_CATEGORY_APPLICATION, "text is \"%s\" \"%s\" %d %d\n", text, composition, cursor, selection_len);
+			}
+			break;
+
+		case SDL_EVENT_MOUSE_MOTION:
+			if (exclusive) {
+				s->rx += e.motion.xrel * m;
+				if (INVERT_MOUSE) {
+					s->ry += e.motion.yrel * m;
+				} else {
+					s->ry -= e.motion.yrel * m;
+				}
+				if (s->rx < 0) {
+					s->rx += RADIANS(360);
+				}
+				if (s->rx >= RADIANS(360)){
+					s->rx -= RADIANS(360);
+				}
+				s->ry = MAX(s->ry, -RADIANS(90));
+				s->ry = MIN(s->ry, RADIANS(90));
+			}
+			break;
+
+		case SDL_EVENT_MOUSE_BUTTON_DOWN:
+			if (e.button.button == SDL_BUTTON_LEFT) {
+				if (exclusive) {
+					if (control) {
+						on_right_click();
+					} else {
+						on_left_click();
+					}
+				} else {
+					SDL_SetRelativeMouseMode(SDL_TRUE);
+				}
+			} else if (e.button.button == SDL_BUTTON_RIGHT) {
+				if (exclusive) {
+					if (control) {
+						on_light();
+					} else {
+						on_right_click();
+					}
+				}
+			} else if (e.button.button == SDL_BUTTON_MIDDLE) {
+				if (exclusive) {
+					on_middle_click();
+				}
+			}
+
+			break;
+
+		case SDL_EVENT_MOUSE_WHEEL:
+			// TODO might have to change this to force 1 step
+			if (e.wheel.direction == SDL_MOUSEWHEEL_NORMAL) {
+				g->item_index += e.wheel.y;
+			} else {
+				g->item_index -= e.wheel.y;
+			}
+			if (g->item_index < 0)
+				g->item_index = item_count -1;
+			else
+				g->item_index %= item_count;
+			break;
+        case SDL_EVENT_WINDOW_RESIZED: {
+            g->scale = get_scale_factor();
+            SDL_GetWindowSizeInPixels(g->window, &g->width, &g->height);
+        }
+        case SDL_EVENT_WINDOW_SHOWN: {
+            // trigger scale calculation when app loads
+            if (first_event) {
+                g->scale = get_scale_factor();
+                SDL_GetWindowSizeInPixels(g->window, &g->width, &g->height);
+                first_event = 0;
+            }
+        }
+		} // switch
+	}
+
+	const Uint8 *state = SDL_GetKeyboardState(NULL);
+
+	if (!g->typing) {
+		float m = dt * 1.0;  // what is this for?
+		g->ortho = state[KEY_ORTHO] ? 64 : 0;
+		g->fov = state[KEY_ZOOM] ? 15 : 65;
+		if (state[KEY_FORWARD]) sz--;
+		if (state[KEY_BACKWARD]) sz++;
+		if (state[KEY_LEFT]) sx--;
+		if (state[KEY_RIGHT]) sx++;
+		if (state[SDL_SCANCODE_LEFT]) s->rx -= m;
+		if (state[SDL_SCANCODE_RIGHT]) s->rx += m;
+		if (state[SDL_SCANCODE_UP]) s->ry += m;
+		if (state[SDL_SCANCODE_DOWN]) s->ry -= m;
+	}
+	float vx, vy, vz;
+	get_motion_vector(g->flying, sz, sx, s->rx, s->ry, &vx, &vy, &vz);
+	if (!g->typing) {
+		if (state[KEY_JUMP]) {
+			if (g->flying) {
+				vy = 1;
+			}
+			else if (dy == 0) {
+				dy = 8;
+			}
+		}
+	}
+	float speed = g->flying ? 20 : 5;
+	int estimate = SDL_roundf(SDL_sqrtf(
+		SDL_powf(vx * speed, 2) +
+		SDL_powf(vy * speed + ABS(dy) * 2, 2) +
+		SDL_powf(vz * speed, 2)) * dt * 8);
+	int step = MAX(8, estimate);
+	float ut = dt / step;
+	vx = vx * ut * speed;
+	vy = vy * ut * speed;
+	vz = vz * ut * speed;
+	for (int i = 0; i < step; i++) {
+		if (g->flying) {
+			dy = 0;
+		}
+		else {
+			dy -= ut * 25;
+			dy = MAX(dy, -250);
+		}
+		s->x += vx;
+		s->y += vy + dy * ut;
+		s->z += vz;
+		if (collide(2, &s->x, &s->y, &s->z)) {
+			dy = 0;
+		}
+	}
+	if (s->y < 0) {
+		s->y = highest_block(s->x, s->z) + 2;
+	}
+
+	return 0;
+
+} // handle_events
+
+void craft::create_window_and_context() {
+    g->start_ticks = (int) SDL_GetTicks(); 
+    Uint32 window_flags = SDL_WINDOW_OPENGL | SDL_WINDOW_HIDDEN;
+    int window_width = WINDOW_WIDTH;
+    int window_height = WINDOW_HEIGHT;
+    if (FULLSCREEN) {
+        SDL_DisplayID display = SDL_GetPrimaryDisplay();
+        int num_modes = 0;
+        const SDL_DisplayMode **modes = SDL_GetFullscreenDisplayModes(display, &num_modes);
+        if (modes) {
+            for (int i = 0; i < num_modes; ++i) {
+                const SDL_DisplayMode *mode = modes[i];
+#if defined(DEBUGGING)
+                SDL_Log("Display %" SDL_PRIu32 " mode %d: %dx%d@%gx %gHz\n", display, i, mode->w, mode->h, mode->pixel_density, mode->refresh_rate);
+#endif
+            }
+        }
+        
+        window_width = modes[num_modes - 1]->w;
+        window_height = modes[num_modes - 1]->h;
+
+        SDL_free(modes);
+
+        window_flags |= SDL_WINDOW_FULLSCREEN;
+    } else {
+        window_flags |= SDL_WINDOW_RESIZABLE;
+    }
+
+    // GL 3.0 + GLSL 130
+    const char* glsl_version = "#version 130";
+#if defined(DEBUGGING)
+    SDL_GL_SetAttribute(SDL_GL_CONTEXT_FLAGS, SDL_GL_CONTEXT_DEBUG_FLAG);
+#else
+    SDL_GL_SetAttribute(SDL_GL_CONTEXT_PROFILE_MASK, SDL_GL_CONTEXT_PROFILE_CORE);
+#endif // DEBUGGING
+    SDL_GL_SetAttribute(SDL_GL_CONTEXT_MAJOR_VERSION, 3);
+    SDL_GL_SetAttribute(SDL_GL_CONTEXT_MINOR_VERSION, 0);
+    SDL_GL_SetAttribute(SDL_GL_DOUBLEBUFFER, 1);
+    SDL_GL_SetAttribute(SDL_GL_DEPTH_SIZE, 24);
+    SDL_GL_SetAttribute(SDL_GL_STENCIL_SIZE, 8);
+
+    g->window = SDL_CreateWindow(this->window_name.c_str(), window_width, window_height, window_flags);
+    g->context = (SDL_GLContext*) SDL_GL_CreateContext(g->window);
+
+    SDL_GL_MakeCurrent(g->window, g->context);
+
+    SDL_GL_SetSwapInterval(1);
+} // create_window_and_context
+
+void craft::parse_buffer(char *buffer) {
+    Player *me = g->players;
+    State *s = &g->players->state;
+    char *key;
+    char *line = tokenize(buffer, "\n", &key);
+    while (line) {
+        int pid;
+        float ux, uy, uz, urx, ury;
+        if (sscanf(line, "U,%d,%f,%f,%f,%f,%f",
+            &pid, &ux, &uy, &uz, &urx, &ury) == 6)
+        {
+            me->id = pid;
+            s->x = ux; s->y = uy; s->z = uz; s->rx = urx; s->ry = ury;
+            force_chunks(me);
+            if (uy == 0) {
+                s->y = highest_block(s->x, s->z) + 2;
+            }
+        }
+        int bp, bq, bx, by, bz, bw;
+        if (sscanf(line, "B,%d,%d,%d,%d,%d,%d",
+            &bp, &bq, &bx, &by, &bz, &bw) == 6)
+        {
+            _set_block(bp, bq, bx, by, bz, bw, 0);
+            if (player_intersects_block(2, s->x, s->y, s->z, bx, by, bz)) {
+                s->y = highest_block(s->x, s->z) + 2;
+            }
+        }
+        if (sscanf(line, "L,%d,%d,%d,%d,%d,%d",
+            &bp, &bq, &bx, &by, &bz, &bw) == 6)
+        {
+            set_light(bp, bq, bx, by, bz, bw);
+        }
+        float px, py, pz, prx, pry;
+        if (sscanf(line, "P,%d,%f,%f,%f,%f,%f",
+            &pid, &px, &py, &pz, &prx, &pry) == 6)
+        {
+            Player *player = find_player(pid);
+            if (!player && g->player_count < MAX_PLAYERS) {
+                player = g->players + g->player_count;
+                g->player_count++;
+                player->id = pid;
+                player->buffer = 0;
+                snprintf(player->name, MAX_NAME_LENGTH, "player%d", pid);
+                update_player(player, px, py, pz, prx, pry, 1); // twice
+            }
+            if (player) {
+                update_player(player, px, py, pz, prx, pry, 1);
+            }
+        }
+        if (sscanf(line, "D,%d", &pid) == 1) {
+            delete_player(pid);
+        }
+        int kp, kq, kk;
+        if (sscanf(line, "K,%d,%d,%d", &kp, &kq, &kk) == 3) {
+            db_set_key(kp, kq, kk);
+        }
+        if (sscanf(line, "R,%d,%d", &kp, &kq) == 2) {
+            Chunk *chunk = find_chunk(kp, kq);
+            if (chunk) {
+                dirty_chunk(chunk);
+            }
+        }
+        double elapsed;
+        int day_length;
+        if (sscanf(line, "E,%lf,%d", &elapsed, &day_length) == 2) {
+            // glfwSetTime(fmod(elapsed, day_length));
+            g->start_time = SDL_fmod(elapsed, day_length) / 1000.0;
+            g->day_length = day_length;
+            g->time_changed = 1;
+        }
+        if (line[0] == 'T' && line[1] == ',') {
+            char *text = line + 2;
+            add_message(text);
+        }
+        char format[64];
+        snprintf(
+            format, sizeof(format), "N,%%d,%%%ds", MAX_NAME_LENGTH - 1);
+        char name[MAX_NAME_LENGTH];
+        if (sscanf(line, format, &pid, name) == 2) {
+            Player *player = find_player(pid);
+            if (player) {
+                strncpy(player->name, name, MAX_NAME_LENGTH);
+            }
+        }
+        snprintf(
+            format, sizeof(format),
+            "S,%%d,%%d,%%d,%%d,%%d,%%d,%%%d[^\n]", MAX_SIGN_LENGTH - 1);
+        int face;
+        char text[MAX_SIGN_LENGTH] = {0};
+        if (sscanf(line, format,
+            &bp, &bq, &bx, &by, &bz, &face, text) >= 6)
+        {
+            _set_sign(bp, bq, bx, by, bz, face, text, 0);
+        }
+        line = tokenize(NULL, "\n", &key);
+    }
+}
+
+void craft::reset_model() {
+    memset(g->chunks, 0, sizeof(Chunk) * MAX_CHUNKS);
+    g->chunk_count = 0;
+    memset(g->players, 0, sizeof(Player) * MAX_PLAYERS);
+    g->player_count = 0;
+    g->observe1 = 0;
+    g->observe2 = 0;
+    g->flying = 0;
+    g->item_index = 0;
+    memset(g->typing_buffer, 0, sizeof(char) * MAX_TEXT_LENGTH);
+    g->typing = 0;
+    memset(g->messages, 0, sizeof(char) * MAX_MESSAGES * MAX_TEXT_LENGTH);
+    g->message_index = 0;
+    g->day_length = DAY_LENGTH;
+    g->start_time = (g->day_length / 3)*1000;
+    // maybe set start_ticks here?
+    g->time_changed = 1;
+}
+
+bool craft::run() {
+    // INITIALIZATION //
+    // curl_global_init(CURL_GLOBAL_DEFAULT);
+    srand(time(NULL));
+    rand();
+
+    // SDL INITIALIZATION //
+    if (SDL_Init(SDL_INIT_EVENTS | SDL_INIT_VIDEO | SDL_INIT_TIMER) < 0) {
+        SDL_LogError(SDL_LOG_CATEGORY_ERROR, "SDL_Init failed (%s)\n", SDL_GetError());
+        return -1;
+    }
+
+    create_window_and_context();
+    if (!g->window) {
+        SDL_LogError(SDL_LOG_CATEGORY_ERROR, "SDL_Window failed (%s)\n", SDL_GetError());
+        SDL_Quit();
+        return -1;
+    }
+
+    SDL_ShowWindow(g->window);
+    SDL_SetRelativeMouseMode(SDL_TRUE);
+
+    if (!gladLoadGLLoader((GLADloadproc) SDL_GL_GetProcAddress)) {
+        SDL_LogError(SDL_LOG_CATEGORY_APPLICATION, 
+            "OpenGL loader failed (%s)\n", SDL_GetError());
+        SDL_Quit();
+        return -1;
+    }
+
+    dump_opengl_info(DUMP_GL_EXTENSIONS);
+
+    glEnable(GL_CULL_FACE);
+    glEnable(GL_DEPTH_TEST);
+    glLogicOp(GL_INVERT);
+    glClearColor(0, 0, 0, 1);
+
+    // LOAD TEXTURES //
+    GLuint texture;
+    glGenTextures(1, &texture);
+    glActiveTexture(GL_TEXTURE0);
+    glBindTexture(GL_TEXTURE_2D, texture);
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
+    load_png_texture("textures/texture.png");
+
+    GLuint font;
+    glGenTextures(1, &font);
+    glActiveTexture(GL_TEXTURE1);
+    glBindTexture(GL_TEXTURE_2D, font);
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
+    load_png_texture("textures/font.png");
+
+    GLuint sky;
+    glGenTextures(1, &sky);
+    glActiveTexture(GL_TEXTURE2);
+    glBindTexture(GL_TEXTURE_2D, sky);
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
+    load_png_texture("textures/sky.png");
+
+    GLuint sign;
+    glGenTextures(1, &sign);
+    glActiveTexture(GL_TEXTURE3);
+    glBindTexture(GL_TEXTURE_2D, sign);
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
+    load_png_texture("textures/sign.png");
+
+    // LOAD SHADERS //
+    Attrib block_attrib = {0};
+    Attrib line_attrib = {0};
+    Attrib text_attrib = {0};
+    Attrib sky_attrib = {0};
+    GLuint program;
+
+    program = load_program(
+        "shaders/block_vertex.glsl", "shaders/block_fragment.glsl");
+    block_attrib.program = program;
+    block_attrib.position = glGetAttribLocation(program, "position");
+    block_attrib.normal = glGetAttribLocation(program, "normal");
+    block_attrib.uv = glGetAttribLocation(program, "uv");
+    block_attrib.matrix = glGetUniformLocation(program, "matrix");
+    block_attrib.sampler = glGetUniformLocation(program, "sampler");
+    block_attrib.extra1 = glGetUniformLocation(program, "sky_sampler");
+    block_attrib.extra2 = glGetUniformLocation(program, "daylight");
+    block_attrib.extra3 = glGetUniformLocation(program, "fog_distance");
+    block_attrib.extra4 = glGetUniformLocation(program, "ortho");
+    block_attrib.camera = glGetUniformLocation(program, "camera");
+    block_attrib.timer = glGetUniformLocation(program, "timer");
+
+    program = load_program(
+        "shaders/line_vertex.glsl", "shaders/line_fragment.glsl");
+    line_attrib.program = program;
+    line_attrib.position = glGetAttribLocation(program, "position");
+    line_attrib.matrix = glGetUniformLocation(program, "matrix");
+
+    program = load_program(
+        "shaders/text_vertex.glsl", "shaders/text_fragment.glsl");
+    text_attrib.program = program;
+    text_attrib.position = glGetAttribLocation(program, "position");
+    text_attrib.uv = glGetAttribLocation(program, "uv");
+    text_attrib.matrix = glGetUniformLocation(program, "matrix");
+    text_attrib.sampler = glGetUniformLocation(program, "sampler");
+    text_attrib.extra1 = glGetUniformLocation(program, "is_sign");
+
+    program = load_program(
+        "shaders/sky_vertex.glsl", "shaders/sky_fragment.glsl");
+    sky_attrib.program = program;
+    sky_attrib.position = glGetAttribLocation(program, "position");
+    sky_attrib.normal = glGetAttribLocation(program, "normal");
+    sky_attrib.uv = glGetAttribLocation(program, "uv");
+    sky_attrib.matrix = glGetUniformLocation(program, "matrix");
+    sky_attrib.sampler = glGetUniformLocation(program, "sampler");
+    sky_attrib.timer = glGetUniformLocation(program, "timer");
+
+    // CHECK COMMAND LINE ARGUMENTS //
+    // if (argc == 2 || argc == 3) {
+    //     g->mode = MODE_ONLINE;
+    //     strncpy(g->server_addr, argv[1], MAX_ADDR_LENGTH);
+    //     g->server_port = argc == 3 ? atoi(argv[2]) : DEFAULT_PORT;
+    //     snprintf(g->db_path, MAX_PATH_LENGTH,
+    //         "cache.%s.%d.db", g->server_addr, g->server_port);
+    // }
+    // else {
+        g->mode = MODE_OFFLINE;
+        snprintf(g->db_path, MAX_PATH_LENGTH, "%s", DB_PATH);
+    // }
+
+    g->create_radius = CREATE_CHUNK_RADIUS;
+    g->render_radius = RENDER_CHUNK_RADIUS;
+    g->delete_radius = DELETE_CHUNK_RADIUS;
+    g->sign_radius = RENDER_SIGN_RADIUS;
+
+    // INITIALIZE WORKER THREADS
+    for (int i = 0; i < WORKERS; i++) {
+        Worker *worker = g->workers + i;
+        worker->index = i;
+        worker->state = WORKER_IDLE;
+        mtx_init(&worker->mtx, mtx_plain);
+        cnd_init(&worker->cnd);
+        thrd_create(&worker->thrd, (thrd_start_t) &worker_run, worker);
+    }
+
+    // DEAR IMGUI INIT //
+    // Setup Dear ImGui context
+    IMGUI_CHECKVERSION();
+    ImGui::CreateContext();
+    ImGuiIO& io = ImGui::GetIO(); (void)io;
+    io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard;
+    io.ConfigFlags |= ImGuiConfigFlags_NavEnableGamepad;
+
+    // Setup Dear ImGui style
+    ImGui::StyleColorsDark();
+    //ImGui::StyleColorsLight();
+
+    // Setup Platform/Renderer backends
+    ImGui_ImplSDL3_InitForOpenGL(g->window, g->context);
+    const char* glsl_version = "#version 130";
+    ImGui_ImplOpenGL3_Init(glsl_version);
+
+    // Load Fonts
+    // - If no fonts are loaded, dear imgui will use the default font. You can also load multiple fonts and use ImGui::PushFont()/PopFont() to select them.
+    // - AddFontFromFileTTF() will return the ImFont* so you can store it if you need to select the font among multiple.
+    // - If the file cannot be loaded, the function will return a nullptr. Please handle those errors in your application (e.g. use an assertion, or display an error and quit).
+    // - The fonts will be rasterized at a given size (w/ oversampling) and stored into a texture when calling ImFontAtlas::Build()/GetTexDataAsXXXX(), which ImGui_ImplXXXX_NewFrame below will call.
+    // - Use '#define IMGUI_ENABLE_FREETYPE' in your imconfig file to use Freetype for higher quality font rendering.
+    // - Read 'docs/FONTS.md' for more instructions and details.
+    // - Remember that in C/C++ if you want to include a backslash \ in a string literal you need to write a double backslash \\ !
+    // - Our Emscripten build process allows embedding fonts to be accessible at runtime from the "fonts/" folder. See Makefile.emscripten for details.
+    //io.Fonts->AddFontDefault();
+    //io.Fonts->AddFontFromFileTTF("c:\\Windows\\Fonts\\segoeui.ttf", 18.0f);
+    //io.Fonts->AddFontFromFileTTF("../../misc/fonts/DroidSans.ttf", 16.0f);
+    //io.Fonts->AddFontFromFileTTF("../../misc/fonts/Roboto-Medium.ttf", 16.0f);
+    //io.Fonts->AddFontFromFileTTF("../../misc/fonts/Cousine-Regular.ttf", 15.0f);
+    //ImFont* font = io.Fonts->AddFontFromFileTTF("c:\\Windows\\Fonts\\ArialUni.ttf", 18.0f, nullptr, io.Fonts->GetGlyphRangesJapanese());
+    //IM_ASSERT(font != nullptr);
+
+    // Our state
+    bool show_demo_window = true;
+    bool show_another_window = false;
+    ImVec4 clear_color = ImVec4(0.45f, 0.55f, 0.60f, 1.00f);
+
+    // RENDER LOOP //
+    int running = 1;
+    while (running) {
+
+        // DATABASE INITIALIZATION //
+        if (g->mode == MODE_OFFLINE || USE_CACHE) {
+            db_enable();
+            if (db_init(g->db_path)) {
+                return -1;
+            }
+            if (g->mode == MODE_ONLINE) {
+                // TODO: support proper caching of signs (handle deletions)
+                db_delete_all_signs();
+            }
+        }
+
+        // CLIENT INITIALIZATION //
+        if (g->mode == MODE_ONLINE) {
+            client_enable();
+            client_connect(g->server_addr, g->server_port);
+            client_start();
+            client_version(1);
+            login();
+        }
+
+        // LOCAL VARIABLES //
+        reset_model();
+        FPS fps = {0, 0, 0};
+        double last_commit = SDL_GetTicks();
+        double last_update = SDL_GetTicks();
+        
+        GLuint sky_buffer = gen_sky_buffer();
+        
+        Player *me = g->players;
+        State *s = &g->players->state;
+        me->id = 0;
+        me->name[0] = '\0';
+        me->buffer = 0;
+        g->player_count = 1;
+
+        // LOAD STATE FROM DATABASE //
+        int loaded = db_load_state(&s->x, &s->y, &s->z, &s->rx, &s->ry);
+        force_chunks(me);
+        if (!loaded) {
+            s->y = highest_block(s->x, s->z) + 2;
+        }
+
+        // BEGIN EVENT LOOP //
+        int previous = SDL_GetTicks();
+        while (1) {
+            glViewport(0, 0, g->width, g->height);
+            // FRAME RATE //
+            if (g->time_changed) {
+                g->time_changed = 0;
+                last_commit = SDL_GetTicks();
+                last_update = SDL_GetTicks();
+                memset(&fps, 0, sizeof(fps));
+            }
+            update_fps(&fps);
+            double now = SDL_GetTicks();
+            double dt = (now - previous) / 1000.0;
+            dt = MIN(dt, 0.2);
+            dt = MAX(dt, 0.0);
+            previous = now;
+#if defined(DEBUGGING)
+            SDL_Log("fps: %d\n", fps.fps);
+#endif
+
+            if (handle_events(dt)) {
+                running = 0;
+                break;
+            }
+            if (g->mode_changed) {
+                g->mode_changed = 0;
+                break;
+            }
+
+        // Start the Dear ImGui frame
+        ImGui_ImplOpenGL3_NewFrame();
+        ImGui_ImplSDL3_NewFrame();
+        ImGui::NewFrame();
+
+        // 1. Show the big demo window (Most of the sample code is in ImGui::ShowDemoWindow()! You can browse its code to learn more about Dear ImGui!).
+        if (show_demo_window)
+            ImGui::ShowDemoWindow(&show_demo_window);
+
+        // 2. Show a simple window that we create ourselves. We use a Begin/End pair to create a named window.
+        {
+            static float f = 0.0f;
+            static int counter = 0;
+
+            ImGui::Begin("Hello, world!");                          // Create a window called "Hello, world!" and append into it.
+
+            ImGui::Text("This is some useful text.");               // Display some text (you can use a format strings too)
+            ImGui::Checkbox("Demo Window", &show_demo_window);      // Edit bools storing our window open/close state
+            ImGui::Checkbox("Another Window", &show_another_window);
+
+            ImGui::SliderFloat("float", &f, 0.0f, 1.0f);            // Edit 1 float using a slider from 0.0f to 1.0f
+            ImGui::ColorEdit3("clear color", (float*)&clear_color); // Edit 3 floats representing a color
+
+            if (ImGui::Button("Button"))                            // Buttons return true when clicked (most widgets return true when edited/activated)
+                counter++;
+            ImGui::SameLine();
+            ImGui::Text("counter = %d", counter);
+
+            ImGui::Text("Application average %.3f ms/frame (%.1f FPS)", 1000.0f / io.Framerate, io.Framerate);
+            ImGui::End();
+        }
+
+        // 3. Show another simple window.
+        if (show_another_window)
+        {
+            ImGui::Begin("Another Window", &show_another_window);   // Pass a pointer to our bool variable (the window will have a closing button that will clear the bool when clicked)
+            ImGui::Text("Hello from another window!");
+            if (ImGui::Button("Close Me"))
+                show_another_window = false;
+            ImGui::End();
+        }
+
+            // HANDLE DATA FROM SERVER //
+            char *buffer = client_recv();
+            if (buffer) {
+                parse_buffer(buffer);
+                free(buffer);
+            }
+
+            // FLUSH DATABASE //
+            if (now - last_commit > COMMIT_INTERVAL) {
+                last_commit = now;
+                db_commit();
+            }
+
+            // SEND POSITION TO SERVER //
+            if (now - last_update > 0.1) {
+                last_update = now;
+                client_position(s->x, s->y, s->z, s->rx, s->ry);
+            }
+
+            // PREPARE TO RENDER //
+            g->observe1 = g->observe1 % g->player_count;
+            g->observe2 = g->observe2 % g->player_count;
+        
+            delete_chunks();
+            del_buffer(me->buffer);
+            
+            me->buffer = gen_player_buffer(s->x, s->y, s->z, s->rx, s->ry);
+            for (int i = 1; i < g->player_count; i++) {
+                interpolate_player(g->players + i);
+            }
+            Player *player = g->players + g->observe1;
+            
+            // RENDER 3-D SCENE //
+            ImGui::Render();
+
+            glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
+            
+            render_sky(&sky_attrib, player, sky_buffer);
+            glClear(GL_DEPTH_BUFFER_BIT);
+            int face_count = render_chunks(&block_attrib, player);
+            render_signs(&text_attrib, player);
+            render_sign(&text_attrib, player);
+            render_players(&block_attrib, player);
+            if (SHOW_WIREFRAME) {
+                render_wireframe(&line_attrib, player);
+            }
+            
+            // RENDER HUD //
+            glClear(GL_DEPTH_BUFFER_BIT);
+            if (SHOW_CROSSHAIRS) {
+                render_crosshairs(&line_attrib);
+            }
+            if (SHOW_ITEM) {
+                render_item(&block_attrib);
+            }
+
+            // RENDER TEXT //
+            char text_buffer[1024];
+            float ts = 12 * g->scale;
+            float tx = ts / 2;
+            float ty = g->height - ts;
+            if (SHOW_INFO_TEXT) {
+                int hour = time_of_day() * 24;
+                char am_pm = hour < 12 ? 'a' : 'p';
+                hour = hour % 12;
+                hour = hour ? hour : 12;
+                snprintf(
+                    text_buffer, 1024,
+                    "(%d, %d) (%.2f, %.2f, %.2f) [%d, %d, %d] %d%cm %dfps",
+                    chunked(s->x), chunked(s->z), s->x, s->y, s->z,
+                    g->player_count, g->chunk_count,
+                    face_count * 2, hour, am_pm, fps.fps);
+                render_text(&text_attrib, ALIGN_LEFT, tx, ty, ts, text_buffer);
+                ty -= ts * 2;
+            }
+            if (SHOW_CHAT_TEXT) {
+                for (int i = 0; i < MAX_MESSAGES; i++) {
+                    int index = (g->message_index + i) % MAX_MESSAGES;
+                    if (strlen(g->messages[index])) {
+                        render_text(&text_attrib, ALIGN_LEFT, tx, ty, ts,
+                            g->messages[index]);
+                        ty -= ts * 2;
+                    }
+                }
+            }
+            if (g->typing) {
+                snprintf(text_buffer, 1024, "> %s", g->typing_buffer);
+                render_text(&text_attrib, ALIGN_LEFT, tx, ty, ts, text_buffer);
+                ty -= ts * 2;
+            }
+            if (SHOW_PLAYER_NAMES) {
+                if (player != me) {
+                    
+                    render_text(&text_attrib, ALIGN_CENTER,
+                        g->width / 2, ts, ts, player->name);
+                    
+                }
+                Player *other = player_crosshair(player);
+                if (other) {
+                    render_text(&text_attrib, ALIGN_CENTER,
+                        g->width / 2, g->height / 2 - ts - 24, ts,
+                        other->name);
+                }
+            }
+
+            // RENDER PICTURE IN PICTURE //
+            if (g->observe2) {
+                player = g->players + g->observe2;
+
+                int pw = 256 * g->scale;
+                int ph = 256 * g->scale;
+                int offset = 32 * g->scale;
+                int pad = 3 * g->scale;
+                int sw = pw + pad * 2;
+                int sh = ph + pad * 2;
+
+                glEnable(GL_SCISSOR_TEST);
+                glScissor(g->width - sw - offset + pad, offset - pad, sw, sh);
+                glClear(GL_COLOR_BUFFER_BIT);
+                glDisable(GL_SCISSOR_TEST);
+                glClear(GL_DEPTH_BUFFER_BIT);
+                glViewport(g->width - pw - offset, offset, pw, ph);
+
+                g->width = pw;
+                g->height = ph;
+                g->ortho = 0;
+                g->fov = 65;
+
+                
+                render_sky(&sky_attrib, player, sky_buffer);
+                
+                glClear(GL_DEPTH_BUFFER_BIT);
+                render_chunks(&block_attrib, player);
+                
+                render_signs(&text_attrib, player);
+                
+                render_players(&block_attrib, player);
+                
+                glClear(GL_DEPTH_BUFFER_BIT);
+                
+                if (SHOW_PLAYER_NAMES) {
+                    
+                    render_text(&text_attrib, ALIGN_CENTER,
+                        pw / 2, ts, ts, player->name);
+                    
+                }
+            }
+            ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
+            SDL_GL_SwapWindow(g->window);
+#if defined(DEBUGGING)
+            gl_check_for_error();
+#endif
+        } // MAIN LOOP
+
+        // SHUTDOWN //
+        db_save_state(s->x, s->y, s->z, s->rx, s->ry);
+        db_close();
+        db_disable();
+        client_stop();
+        client_disable();
+        del_buffer(sky_buffer);
+        delete_all_chunks();
+        delete_all_players();
+    } // RENDER LOOP
+
+    SDL_Log("Cleaning up ImGui objects. . .");
+    ImGui_ImplOpenGL3_Shutdown();
+    ImGui_ImplSDL3_Shutdown();
+    ImGui::DestroyContext();
+
+    SDL_Log("Cleaning up OpenGL objects. . .");
+    glDeleteTextures(1, &texture);
+    glDeleteTextures(1, &font);
+    glDeleteTextures(1, &sky);
+    glDeleteTextures(1, &sign);
+    glDeleteProgram(block_attrib.program);
+    glDeleteProgram(text_attrib.program);
+    glDeleteProgram(sky_attrib.program);
+    glDeleteProgram(line_attrib.program);
+#if defined(DEBUGGING)
+    gl_check_for_error();
+#endif
+
+    SDL_Log("Cleaning up SDL objects. . .");
+    SDL_GL_DeleteContext(g->context);
+    SDL_DestroyWindow(g->window);
+    SDL_Quit();
+    
+    // curl_global_cleanup();
+
+    return true;
+} // run
