+++ sources/imaze.h
@@ -1,0 +1,9 @@
+#ifndef IMAZE_H
+#define IMAZE_H
+
+class imaze {
+public:
+    virtual bool build(unsigned int seed = 0) = 0;
+};
+
+#endif // IMAZE_H
